==> src/01_get_retro_game.R <==
##
## Aggregate data from Retrosheet
## https://www.retrosheet.org/gamelogs/index.html
##

library(tidyverse)
library(lubridate)
library(glue)

# See https://www.retrosheet.org/gamelogs/glfields.txt.
data_dictionary <- tribble(
  ~col_name, ~col_type, ~notes,
  "game_dt", "c", NA_character_,
  "game_num_of_dt", "c", "0=single, 1=first game of doubleheader, ...",
  "day_of_wk", "c", NA_character_,
  "vis_team", "c", NA_character_,
  "vis_league", "c", NA_character_,
  "vis_game_num", "c", NA_character_,
  "home_team", "c", NA_character_,
  "home_league", "c", NA_character_,
  "home_game_num", "c", NA_character_,
  "vis_R", "d", NA_character_,
  "home_R", "d", NA_character_,
  "outs", "d", NA_character_,
  "day_night", "c", NA_character_,
  "completion_info", "c", "if completed late, yyyymmdd,park,vs,hs,len",
  "forfeit", "c", "V: visitor forfeited, H: home forfeited, T: no-decision",
  "protest", "c", NA_character_,
  "park_id", "c", NA_character_,
  "attendance", "d", NA_character_,
  "G_min", "d", NA_character_,
  "vis_line_score", "c", NA_character_,
  "home_line_score", "c", NA_character_,
  "vis_AB", "d", NA_character_,
  "vis_H", "d", NA_character_,
  "vis_2B", "d", NA_character_,
  "vis_3B", "d", NA_character_,
  "vis_HR", "d", NA_character_,
  "vis_RBI", "d", NA_character_,
  "vis_sac", "d", NA_character_,
  "vis_sac_flies", "d", NA_character_,
  "vis_HBP", "d", NA_character_,
  "vis_BB", "d", NA_character_,
  "vis_IBB", "d", NA_character_,
  "vis_K", "d", NA_character_,
  "vis_SB", "d", NA_character_,
  "vis_CS", "d", NA_character_,
  "vis_GDP", "d", NA_character_,
  "vis_cint", "d", "awarded first on catcher's interference",
  "vis_LOB", "d", NA_character_,
  "vis_pitchers", "d", NA_character_,
  "vis_ind_ER", "d", "individual earned runs",
  "vis_team_ER", "d", NA_character_,
  "vis_WP", "d", NA_character_,
  "vis_balks", "d", NA_character_,
  "vis_putouts", "d", NA_character_,
  "vis_assists", "d", NA_character_,
  "vis_errors", "d", NA_character_,
  "vis_passed_balls", "d", NA_character_,
  "vis_double_plays", "d", NA_character_,
  "vis_triple_plays", "d", NA_character_,
  "home_AB", "d", NA_character_,
  "home_H", "d", NA_character_,
  "home_2B", "d", NA_character_,
  "home_3B", "d", NA_character_,
  "home_HR", "d", NA_character_,
  "home_RBI", "d", NA_character_,
  "home_sac", "d", NA_character_,
  "home_sac_flies", "d", NA_character_,
  "home_HBP", "d", NA_character_,
  "home_BB", "d", NA_character_,
  "home_IBB", "d", NA_character_,
  "home_K", "d", NA_character_,
  "home_SB", "d", NA_character_,
  "home_CS", "d", NA_character_,
  "home_GDP", "d", NA_character_,
  "home_cint", "d", "awarded first on catcher's interference",
  "home_LOB", "d", NA_character_,
  "home_pitchers", "d", NA_character_,
  "home_ind_ER", "d", "individual earned runs",
  "home_team_ER", "d", NA_character_,
  "home_WP", "d", NA_character_,
  "home_balks", "d", NA_character_,
  "home_putouts", "d", NA_character_,
  "home_assists", "d", NA_character_,
  "home_errors", "d", NA_character_,
  "home_passed_balls", "d", NA_character_,
  "home_double_plays", "d", NA_character_,
  "home_triple_plays", "d", NA_character_,
  "ump_hp_id", "c", NA_character_,
  "ump_hp_name", "c", NA_character_,
  "ump_1b_id", "c", NA_character_,
  "ump_1b_name", "c", NA_character_,
  "ump_2b_id", "c", NA_character_,
  "ump_2b_name", "c", NA_character_,
  "ump_3b_id", "c", NA_character_,
  "ump_3b_name", "c", NA_character_,
  "ump_lf_id", "c", NA_character_,
  "ump_lf_name", "c", NA_character_,
  "ump_rf_id", "c", NA_character_,
  "ump_rf_name", "c", NA_character_,
  "vis_mgr_id", "c", NA_character_,
  "vis_mgr_name", "c", NA_character_,
  "home_mgr_id", "c", NA_character_,
  "home_mgr_name", "c", NA_character_,
  "winning_pitcher_id", "c", NA_character_,
  "winning_pitcher_name", "c", NA_character_,
  "losing_pitcher_id", "c", NA_character_,
  "losing_pitcher_name", "c", NA_character_,
  "saving_pitcher_id", "c", NA_character_,
  "saving_pitcher_name", "c", NA_character_,
  "game_winning_rbi_id", "c", NA_character_,
  "game_winning_rbi_name", "c", NA_character_,
  "vis_starting_p_id", "c", NA_character_,
  "vis_starting_p_name", "c", NA_character_,
  "home_starting_p_id", "c", NA_character_,
  "home_starting_p_name", "c", NA_character_,
  "vis_b1_id", "c", NA_character_,
  "vis_b1_name", "c", NA_character_,
  "vis_b1_pos", "c", NA_character_,
  "vis_b2_id", "c", NA_character_,
  "vis_b2_name", "c", NA_character_,
  "vis_b2_pos", "c", NA_character_,
  "vis_b3_id", "c", NA_character_,
  "vis_b3_name", "c", NA_character_,
  "vis_b3_pos", "c", NA_character_,
  "vis_b4_id", "c", NA_character_,
  "vis_b4_name", "c", NA_character_,
  "vis_b4_pos", "c", NA_character_,
  "vis_b5_id", "c", NA_character_,
  "vis_b5_name", "c", NA_character_,
  "vis_b5_pos", "c", NA_character_,
  "vis_b6_id", "c", NA_character_,
  "vis_b6_name", "c", NA_character_,
  "vis_b6_pos", "c", NA_character_,
  "vis_b7_id", "c", NA_character_,
  "vis_b7_name", "c", NA_character_,
  "vis_b7_pos", "c", NA_character_,
  "vis_b8_id", "c", NA_character_,
  "vis_b8_name", "c", NA_character_,
  "vis_b8_pos", "c", NA_character_,
  "vis_b9_id", "c", NA_character_,
  "vis_b9_name", "c", NA_character_,
  "vis_b9_pos", "c", NA_character_,
  "home_b1_id", "c", NA_character_,
  "home_b1_name", "c", NA_character_,
  "home_b1_pos", "c", NA_character_,
  "home_b2_id", "c", NA_character_,
  "home_b2_name", "c", NA_character_,
  "home_b2_pos", "c", NA_character_,
  "home_b3_id", "c", NA_character_,
  "home_b3_name", "c", NA_character_,
  "home_b3_pos", "c", NA_character_,
  "home_b4_id", "c", NA_character_,
  "home_b4_name", "c", NA_character_,
  "home_b4_pos", "c", NA_character_,
  "home_b5_id", "c", NA_character_,
  "home_b5_name", "c", NA_character_,
  "home_b5_pos", "c", NA_character_,
  "home_b6_id", "c", NA_character_,
  "home_b6_name", "c", NA_character_,
  "home_b6_pos", "c", NA_character_,
  "home_b7_id", "c", NA_character_,
  "home_b7_name", "c", NA_character_,
  "home_b7_pos", "c", NA_character_,
  "home_b8_id", "c", NA_character_,
  "home_b8_name", "c", NA_character_,
  "home_b8_pos", "c", NA_character_,
  "home_b9_id", "c", NA_character_,
  "home_b9_name", "c", NA_character_,
  "home_b9_pos", "c", NA_character_,
  "add_info", "c", NA_character_,
  "acq_info", "c", "Y=incomplete, N=missing, D=derived, P=portion"
)

# Aggregate all the raw data sets into a single data frame.
retro_logs_raw <- map_df(
  list.files("data/retrosheet/game"),
  ~read_csv(
    file.path("data/retrosheet", .x),
    col_names = c(data_dictionary$col_name, paste0("c", (nrow(data_dictionary)+1):161)),
    col_types = paste(c(data_dictionary$col_type, strrep("c", 161-nrow(data_dictionary))), collapse = "")
  )
)

# Clean the data
retro_logs <- retro_logs_raw %>%
  mutate(
    game_dt = ymd(game_dt),
    dur = duration(G_min, units = "minutes")
  )

# Save final data frame.
#
saveRDS(retro_logs, file.path("data", "retro_game.rds"))
