==> src/02_exploration.Rmd <==
---
title: "Cleveland Baseball Game Duration"
subtitle: "Data Exploration"
author: "Michael Foley"
date: "`r Sys.Date()`"
output: 
  html_document:
    theme: united
    highlight: tango
    css: "style.css"
    code_folding: hide
    toc: true
    toc_depth: 2
    toc_float: true
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)

library(tidyverse)
library(lubridate)
library(scales)
library(ggtext)
library(extrafont)
# library(fpp3)
library(tsibble)
library(fable)
# library(patchwork) # for arranging plots
library(glue)
library(png)
library(gridGraphics)
library(grid)
library(skimr)
library(ggrepel)
```

```{r data}
# Color schemes, utility functions.
source("00_setup.R")

# Experimenting with alternative markers on plots.
ball_png <- png::readPNG("../assets/baseball.png", native = TRUE)
ball_grob <- grid::rasterGrob(ball_png, interpolate = FALSE)

fig_no <- 0
tbl_no <- 0

# Refresh 2023 data?
refresh_2023 <- FALSE

flextable::set_flextable_defaults(
  theme_fun = "theme_vanilla",
  font.family = "Tahoma",
  font.size = 10,
  font.color = "#333333"
)
```

Average pace of play slowed from a little over two hours per nine-innings in the 1940s to a brutal 3:16 in 2021. This project investigates reasons for the increase, and measures the effect of recent efforts to quicken the pace of play. This section is an overview of these data sources and what they offer for analysis. 

- Cleveland game and season summaries scraped from [Baseball-Reference](https://www.baseball-reference.com/teams/CLE/2023.shtml) web pages.
- Team season summaries and MLB pace of play from MLB's publicly accessible (but apparently undocumented) API.
- [Retrosheet's](https://www.retrosheet.org/) game and event files.

In addition to insights that should guide further research, this script performs minor data engineering tasks to prep the data sets for model fitting. The output is four time-series tsibbles, `bref_cle_season_ts`, `mlb_season_ts`, `retro_season_g_ts`, and `retro_season_e_ts` saved in the project [data](https://github.com/mpfoley73/cleveland-mlb-games/tree/main/data) directory.

# Raw Data from Step 01

Prior to running this report, I ran

- [01_get_bref_cle.R](https://github.com/mpfoley73/cleveland-mlb-games/blob/main/src/01_get_bref_cle.R) to scrape game summaries from Baseball-Reference.com. This source updates daily, so re-run to refresh YTD. 
- [01_get_mlb.R](https://github.com/mpfoley73/cleveland-mlb-games/blob/main/src/01_get_mlb.R) to pull from the MLB API. Also updates daily.
- [01-get-retro_game.R](https://github.com/mpfoley73/mlb-games/blob/main/src/01_get_retro_game.R) to load game summaries downloaded from Retrosheet. Updates once per season, so no need to re-run.
- [01-get-retro-even.R](https://github.com/mpfoley73/mlb-games/blob/main/src/01_get_retro_event.R) to download and process with Chadwick game-event files from Retrosheet.  Updates once per season.

```{r echo=TRUE, class.source='fold-show', warning=FALSE, message=FALSE}
# 01_get_bref_cle.R. Re-run source to refresh 2023.
if(refresh_2023) source("01_get_bref_cle.R")
bref_cle_games <- readRDS(file.path(proj_dir, "data/bref_cle_games.rds"))
bref_cle_batting <- readRDS(file.path(proj_dir, "data/bref_cle_batting.rds"))
bref_cle_pitching <- readRDS(file.path(proj_dir, "data/bref_cle_pitching.rds"))

# 01_get_mlb.R. Re-run source to refresh 2023.
last_run <- ymd("2023-07-19")
if(refresh_2023) source("01_get_mlb.R")
mlb_pace <- readRDS(file.path(proj_dir, "data/mlb_pace.rds"))
mlb_batting <- readRDS(file.path(proj_dir, "data/mlb_batting.rds"))
mlb_pitching <- readRDS(file.path(proj_dir, "data/mlb_pitching.rds"))

# 01_get_retro_game.
retro_games <- readRDS(file.path(proj_dir, "data/retro_game.rds"))

# 01_get_retro_event.
retro_events <- readRDS(file.path(proj_dir, "data/retro_event.rds"))

```

<br>
All four sources provide game duration. Baseball-reference was the most difficult to assemble, so I only collected Cleveland games. It does not have any data that isn't available in one or more of the other data sources. The MLB-API is nice because it covers the current season, and includes pitch counts. The Retrosheet Games data was simple to pull, but has limited coverage. The gold standard is the Retrosheet Events data. Difficult to assemble, but it has everything, including mid-inning pitcher substitutions. The only problem is coverage - it only goes back to 1914. I may ultimately use multiple sources in the final analysis.

<br>

```{r}
tribble(
  ~Source, ~Duration, ~Pitchers, ~Pitches, ~`Mid-Inning Subs`, ~Coverage,
  "Baseball-Reference", "Y", "", "", "", "1900-present, Cleveland only.",
  "MLB API", "Y", "Y", "Y", "", "1900-present.",
  "Retrosheet Games", "Y", "Y", "", "", "1900-2022.",
  "Retroshet Events", "Y", "Y", "Y", "Y", "1914-2022."
) %>%
  flextable::flextable() %>%
  flextable::autofit()
```

<font color = "#666666" size = -1><b>Table `r (tbl_no <- tbl_no + 1)`</b>. Coverage for data sources used in this analysis. Baseball-Reference and MLB API last collected on `r max(bref_cle_games$game_date) %>% format("%b %d, %Y")`.</font>

## Baseball-Reference {.tabset .tabset-fade .tabset-pills}

R script **01_get_bref_cle.R** scraped `r n_distinct(bref_cle_games$season)` seasons of Cleveland baseball game summaries, including the 2023 season through `r format(max(bref_cle_games$game_date), "%b %d")`, from [Baseball-Reference](https://www.baseball-reference.com/teams/CLE/2023.shtml). Baseball-Reference has one web page for each team and season. The pages include a table of game summaries. The R script loops through each season and scrapes the table. I ran it through all seasons through 2022, then set it to just refresh the 2023 season. _Note to self: run the script periodically to refresh this analysis by setting_ `refresh_2023 = TRUE`. Data set `bref_cle_games` is a `r comma(nrow(bref_cle_games))` x `r ncol(bref_cle_games)` data frame with one row per each of the `r comma(nrow(bref_cle_games))` games from 1901 - `r format(max(bref_cle_games$game_date), "%b %d, %Y")`. The Baseball-Reference pages also have player summaries which the R script scrapes as well. `bref_cle_batting` (`r comma(nrow(bref_cle_batting))` x `r ncol(bref_cle_batting)`) has season statistics for `r comma(length(unique(bref_cle_batting$Name)), 1)` Cleveland batters. `bref_cle_pitching` (`r comma(nrow(bref_cle_pitching))` x `r ncol(bref_cle_pitching)`) has stats for `r comma(length(unique(bref_cle_pitching$Name)), 1)` pitchers.

### bref_cle_games

```{r}
skimr::skim(bref_cle_games)
```

### bref_cle_batting

```{r}
skimr::skim(bref_cle_batting)
```

### bref_cle_pitching

```{r}
skimr::skim(bref_cle_pitching)
```

## MLB API {.tabset .tabset-fade .tabset-pills}

The MLB API is publicly accessible. R package [baseballr](https://github.com/BillPetti/baseballr/) is an excellent tool for pulling this data. However, the MLB API seems to have no public documentation, and the functionality changes. I had trouble getting **baseballr** to work, probably because the API is a moving target. I was able to copy/paste the important code into my own script, **01_get_mlb.R**, and get the data I needed. One particularly useful module in the API is **gamePace**, an incredible set of metrics related to game pace. You can get data at the individual game level, or summarized by team. I originally wanted to focus on just Cleveland, but the team summaries only show the team's _home_ games. So if you want season summaries for one team, you need game summaries for _all_ teams, then untangle the home and visting data. This was just too much effort, so I shifted my focus to MLB as a whole. As with **01_get_bref_cle.R**, I used my MLB R script download all seasons through 2022, then set it to just refresh 2023. _Run this script too when you refresh 2023!_ Data set `mlb_pace` is a `r comma(nrow(mlb_pace))` x `r ncol(mlb_pace)` data frame with one row per each season, 1901-1923. Data sets `mlb_batting` (`r comma(nrow(mlb_batting))` x `r ncol(mlb_batting)`) and `mlb_pitching` (`r comma(nrow(mlb_pitching))` x `r ncol(mlb_pitching)`) have one row per team and season. I summarized the player summaries to augment the `mlb_pace` data with offense and defense statistics.

### mlb_pace

```{r}
skimr::skim(mlb_pace)
```

### mlb_batting

```{r}
skimr::skim(mlb_batting)
```

### mlb_pitching

```{r}
skimr::skim(mlb_pitching)
```

## Retrosheet {.tabset .tabset-fade .tabset-pills}

[Retrosheet](https://www.retrosheet.org/) is a voluntary organization that has compiled game data into digital form and made it available for public use. Retrosheet has two main types of data. Game logs summarize the main outcomes of individual games, sort of like box scores. This data goes all the back to the 1800s. I downloaded this data through 2022 then processed it with **01_get_retro_game.R**. Event logs summarize individual plays. This data is incredibly detailed, but comes at the cost of having to process and store it in a database. Also, like the game logs, the data is published once per season, so you cannot get year to date data. **01_get_retro_event.R** downloaded and processed this data. Data set `retro_games` (`r nrow(retro_games) %>% comma(1)` x `r ncol(retro_games)`) summarizes games from `r min(year(retro_games$game_dt))` - `r max(year(retro_games$game_dt))`. `retro_events` (`r nrow(retro_events) %>% comma(1)` x `r ncol(retro_events)`) summarizes events into augmented game summaries that include more details that I am interested in correlating with pace of play. The data ranges from `r min(retro_events$year_id)` to `r max(retro_events$year_id)`.

### retro_games

```{r}
skimr::skim(retro_games)
```

### retro_events

```{r}
skimr::skim(retro_events)
```

# Create tsibbles {.tabset .tabset-fade .tabset-pills}

Baseball-Reference has game-level summaries for Cleveland games. MLB API has season-level summaries of MLB. Retrosheet game (g) and Retrosheet Event (e) have game-level summaries for every MLB game. The following chunks summarize the raw data sets into four time series tibbles (tsibbles). Each tsibble has one row per season. Most measures are standardized into per-game (`_G`), per-nine-innings (`_9I`), or per-plate-appearance (`_PA`).

## bref_cle_season_ts

```{r echo=TRUE}
# key includes doubleheader game in order to index by game.
bref_cle_games_ts <- bref_cle_games %>%
  mutate(
    hrs = dur / dhours(1),
    hrs_9I = hrs / innings * 9, # Calc ratio here because hrs is sometimes NA
    R = runs_scored + runs_allowed
  ) %>% 
  tsibble(key = c(doubleheader_game), index = game_date)

bref_cle_batting_ts <- bref_cle_batting %>%
  mutate(`1B` = H - (`2B` + `3B` + HR)) %>% # Singles are used in SLG calc.
  relocate(`1B`, .before = 10) %>%
  summarize(.by = season, across(c(PA:SO, TB:IBB), ~sum(.x, na.rm = TRUE))) %>%
  mutate(
    BA = H / AB,
    OBP = (H + BB + HBP) / (AB + BB + HBP + SF),
    SLG = (`1B` + 2*`2B` + 3*`3B` + 4*HR) / AB,
    OPS = OBP + SLG
  ) %>%
  rename_with(~ paste0("bat_", .x), c(PA:OPS)) %>%
  tsibble(index = season)

bref_cle_pitching_ts <- bref_cle_pitching %>%
  summarize(.by = season, across(c(G:BF), ~sum(.x, na.rm = TRUE))) %>%
  mutate(
    ERA = 9 * ER / IP,
    WHIP = (BB + H) / IP,
  ) %>%
  select(-c(GS, GF, SV)) %>%
  rename_with(~ paste0("pit_", .x), c(G:WHIP)) %>%
  tsibble(index = season)

# `bref_cle_games_ts` is at the game granularity. Summarize it to the season level
# and combine with the batting and pitching season summaries to create a singe 
# tsibble.

bref_cle_season_ts <- bref_cle_games_ts %>%
  arrange(season, game_no) %>% #filter(season == 2020)
  index_by(season = ~ year(.)) %>%
  summarize(
    G = n(),
    I = sum(innings),
    I_G = mean(innings),
    hrs_G = mean(hrs, na.rm = TRUE),
    hrs_9I = mean(hrs_9I, na.rm = TRUE),
    na_dur = sum(is.na(dur)),
    home_G = sum(home_ind == "Home"),
    home_att_G = mean(if_else(home_ind == "Home", attendance, NA_real_), na.rm = TRUE),
    na_home_att = sum(home_ind == "Home" & doubleheader_game == 0 & is.na(attendance)),
    W = sum(str_sub(outcome, 1, 1) == "W"),
    L = sum(str_sub(outcome, 1, 1) == "L"),
    avg_R_9I = mean(R / I),
    place = last(new_rank),
    # home_games = sum(home_ind == "Home"),
    # away_games = sum(home_ind == "Away")
  ) %>%
  mutate(W_pct = W / (W + L)) %>%
  relocate(c(W_pct, place), .after = 6) %>%
  inner_join(bref_cle_batting_ts, join_by(season)) %>%
  inner_join(bref_cle_pitching_ts, join_by(season)) %>%
  mutate(
    PA = bat_PA + pit_BF,
    PA_G = PA / G,
    PA_9I = (bat_PA + pit_BF) / I * 9,
    mins_PA = hrs_G / PA_G * 60, # because hrs may be NA
    R_9I = (bat_R + pit_R) / I * 9,
    R_PA = (bat_R + pit_R) / PA,
    H_9I = (bat_H + pit_H) / I * 9,
    H_PA = (bat_H + pit_H) / PA,
    HR_9I = (bat_HR + pit_HR) / I * 9,
    HR_PA = (bat_HR + pit_HR) / PA,
    BB_9I = (bat_BB + pit_BB) / I * 9,
    BB_PA = (bat_BB + pit_BB) / PA,
    SO_9I = (bat_SO + pit_SO) / I * 9,
    SO_PA = (bat_SO + pit_SO) / PA
  ) %>%
  select(
    season, G, I_G, PA_9I, hrs_9I, mins_PA, R_9I, R_PA, H_9I, H_PA, HR_9I, 
    HR_PA, BB_9I, BB_PA, SO_9I, SO_PA, ERA = pit_ERA, WHIP = pit_WHIP,
    home_G, home_att_G, na_dur, na_home_att,
    W, L, W_pct, place
  )

glimpse(bref_cle_season_ts)
```

## mlb_season_ts_0

```{r echo=TRUE}
mlb_batting_ts <- mlb_batting %>%
  # Singles are used in SLG calc.
  mutate(`1B` = H - (`2B` + `3B` + HR)) %>%
  relocate(`1B`, .before = 6) %>%
  summarize(.by = season, across(c(G:HBP, AB, SB:PA, TB:K), ~sum(.x, na.rm = TRUE))) %>%
  mutate(
    P = if_else(season <= 1973, NA_integer_, P),
    K = if_else(season <= 1909, NA_integer_, K),
    BA = H / AB,
    OBP = (H + BB + HBP) / (AB + BB + HBP + sac_bunts),
    SLG = (`1B` + 2*`2B` + 3*`3B` + 4*HR) / AB,
    OPS = OBP + SLG
  ) %>%
  rename_with(~ paste0("batting_", .x), c(G:OPS)) %>%
  tsibble(index = season)

mlb_pitching_ts <- mlb_pitching %>%
  summarize(.by = season, across(c(G:HBP, AB, IP:ER, BF:wild_pitches), ~sum(.x, na.rm = TRUE))) %>%
  rename_with(~ paste0("pitching_", .x), c(G:wild_pitches)) %>%
  tsibble(index = season)

mlb_pace_ts <- tsibble(mlb_pace, index = season) %>%
  mutate(
    PA_9I = PA / I * 9,
    PA_pitcher = PA / pitchers,
    pitchers_9I = pitchers / I * 9,
    P_9I = P / I * 9,
    P_PA = P / PA,
    P_pitcher = P / pitchers,
    hrs_9I = as.numeric(dur, "hours") / I * 9,
    mins_PA = as.numeric(dur, "minutes") / PA,
    secs_P = as.numeric(dur, "seconds") / P,
    I_G = I / G,
    R_9I = R / I * 9,
    R_PA = R / PA,
    H_9I = H / I * 9,
    H_PA = H / PA
  ) %>%
  select(season, G, I_G, PA_9I:H_PA, 
         everything(), - c(extra_I_dur, extra_I_G))

mlb_season_ts_0 <- mlb_pace_ts %>%
  inner_join(mlb_batting_ts, by = join_by(season)) %>%
  inner_join(mlb_pitching_ts, by = join_by(season)) %>%
  mutate(
    HR_9I = batting_HR / I * 9,
    HR_PA = batting_HR / PA,
    BB_9I = batting_BB / I * 9,
    BB_PA = batting_BB / PA,
    SO_9I = pitching_K / I * 9,
    SO_PA = pitching_K / PA,
    SB_9I = batting_SB / I * 9,
    CS_9I = batting_CS / I * 9,
    ERA = pitching_ER / I * 9 / 2, # two pitchers
    WHIP = (batting_BB + H) / I / 2 # two pitchers
  ) %>%
  select(season:H_PA, HR_9I:last_col())

glimpse(mlb_season_ts_0)
```

## retro_season_g_ts

```{r echo=TRUE}
retro_season_g_ts <- retro_games %>%
  mutate(
    season = year(game_dt),
    I = outs / 6,  # outs are NA in ~500 rows
    # PA = vis_AB + home_AB + vis_BB + home_BB + vis_sac + home_sac,
    # PA := https://www.mlb.com/glossary/standard-stats/plate-appearance
    PA = home_LOB + vis_LOB + outs + home_R + vis_R,
    pitchers = vis_pitchers + home_pitchers
  ) %>%
  summarise(
    .by = season,
    G = n(),
    I_G = sum(I) / G,
    PA_9I = mean(PA / I * 9, na.rm = TRUE),
    PA_pitcher = mean(PA / pitchers, na.rm = TRUE),
    pitchers_9I = mean(pitchers / I * 9, na.rm = TRUE),
    hrs_9I = mean(G_min / 60 / I * 9, na.rm = TRUE),
    mins_PA = mean(G_min / PA, na.rm = TRUE),
    R_9I = mean((vis_R + home_R) / I * 9, na.rm = TRUE),
    R_PA = mean((vis_R + home_R) / PA, na.rm = TRUE),
    H_9I = mean((vis_H + home_H) / I * 9, na.rm = TRUE),
    H_PA = mean((vis_H + home_H) / PA, na.rm = TRUE),
    HR_9I = mean((vis_HR + home_HR) / I * 9, na.rm = TRUE),
    HR_PA = mean((vis_HR + home_HR) / PA, na.rm = TRUE),
    BB_9I = mean((vis_BB + home_BB) / I * 9, na.rm = TRUE),
    BB_PA = mean((vis_BB + home_BB) / PA, na.rm = TRUE),
    SO_9I = mean((vis_K + home_K) / I * 9, na.rm = TRUE),
    SO_PA = mean((vis_K + home_K) / PA, na.rm = TRUE),
    SB_9I = mean((vis_SB + home_SB) / I * 9, na.rm = TRUE),
    CS_9I = mean((vis_CS + home_CS) / I * 9, na.rm = TRUE),
    ERA = mean((vis_team_ER + home_team_ER) / I * 9, na.rm = TRUE),
    WHIP = mean((vis_BB + home_BB + vis_H + home_H) / I, na.rm = TRUE),
    att_G = mean(attendance, na.rm = TRUE),
    na_dur = sum(is.na(G_min)),
    na_att = sum(is.na(attendance))
  ) %>%
  tsibble(index = season)

glimpse(retro_season_g_ts)
```

## retro_season_e_ts

```{r echo=TRUE}
retro_season_e_ts_0 <- retro_events %>%
  rename(season = year_id) %>%
  mutate(
    # PA := https://www.mlb.com/glossary/standard-stats/plate-appearance
    PA = home_lob_ct + away_lob_ct + outs_ct + home_score_ct + away_score_ct,
    pitchers = away_pitcher_ct + home_pitcher_ct,
    P = pitches, # balls + strikes,
    I = outs_ct / 6
  ) %>%
  summarise(
    .by = season,
    G = n(),
    I_G = sum(I) / G,
    PA_9I = mean(PA / I * 9, na.rm = TRUE),
    PA_pitcher = mean(PA / pitchers, na.rm = TRUE),
    pitchers_9I = mean(pitchers / I * 9, na.rm = TRUE),
    pitsubsbi_9I = mean((pitchers - mid_inning_pitcher_sub_ct - 2) / I * 9, na.rm = TRUE),
    pitsubsmi_9I = mean(mid_inning_pitcher_sub_ct / I * 9, na.rm = TRUE),
    P_9I = mean(P / I * 9, na.rm = TRUE),
    P_PA = mean(P / PA, na.rm = TRUE),
    P_pitcher = mean(P / pitchers, na.rm = TRUE),
    hrs_9I = mean(minutes_game_ct / 60 / I * 9, na.rm = TRUE),
    mins_PA = mean(minutes_game_ct / PA, na.rm = TRUE),
    secs_P = mean(minutes_game_ct * 60 / P, na.rm = TRUE),
    R_9I = mean((away_score_ct + home_score_ct) / I * 9, na.rm = TRUE),
    R_PA = mean((away_score_ct + home_score_ct) / PA, na.rm = TRUE),
    H_9I = mean((away_hits_ct + home_hits_ct) / I * 9, na.rm = TRUE),
    H_PA = mean((away_hits_ct + home_hits_ct) / PA, na.rm = TRUE),
    HR_9I = mean((away_hr_ct + home_hr_ct) / I * 9, na.rm = TRUE),
    HR_PA = mean((away_hr_ct + home_hr_ct) / PA, na.rm = TRUE),
    BB_9I = mean((away_bb_ct + home_bb_ct) / I * 9, na.rm = TRUE),
    BB_PA = mean((away_bb_ct + home_bb_ct) / PA, na.rm = TRUE),
    SO_9I = mean((away_so_ct + home_so_ct) / I * 9, na.rm = TRUE),
    SO_PA = mean((away_so_ct + home_so_ct) / PA, na.rm = TRUE),
    SB_9I = mean((away_sb_ct + home_sb_ct) / I * 9, na.rm = TRUE),
    CS_9I = mean((away_cs_ct + home_cs_ct) / I * 9, na.rm = TRUE),
    ERA = mean((away_er_ct + home_er_ct) / I * 9, na.rm = TRUE),
    WHIP = mean((away_bb_ct + home_bb_ct + away_hits_ct + home_hits_ct) / I, na.rm = TRUE),
    att_G = mean(attend_park_ct, na.rm = TRUE),
    na_dur = sum(is.na(minutes_game_ct)),
    na_att = sum(is.na(attend_park_ct))
  ) %>%
  tsibble(index = season)

glimpse(retro_season_e_ts_0)
```

# Coverage and Quality

Most features are fully populated, but there are several exceptions. There are also some data quality issues.

The MLB-API duration data has issues before 1930 and its plate appearances are odd before 1920. Imputation gives more reasonable values. There is no reliable source of pitch counts prior to 1999, so I've delete them.

The pitch count per game (**MLB Pitches**) is potentially among the most important variables, but unfortunately the data is not populate until 1950. And even then, something is very wrong about the values before 1988. Furthermore, the 1988-1998 values are suspiciously noisy. I null out the pre-1999 values for `total_P`.

The stolen bases data does not pass the eyeball test prior to 1950. Stolen bases seem reasonable with the exception of 1914-15. I will leave this alone because I am not sure what to delete or replace. I'll just use stolen bases prior to 1950 with caution.

## Duration {.tabset .tabset-fade .tabset-pills}

```{r}
fig_no <- fig_no + 1
```

The MLB API appears to have the most complete record of game duration (Fig. `r fig_no`a), but this data is pre-summarized at the season level, so 100% coverage since 1900 just means the API returned a sum of minutes for each season. (Fig. `r fig_no`b) shows that when expressed per nine innings, something is quite off for eight of the seasons. One option is to impute values from the other sources, but I will try MICE instead.^[More on how to use MICE in [my blog](https://mpfoley73.netlify.app/post/2022-07-26-imputing-values-with-mice/)] Fig. `r fig_no`d shows the imputation results, and Fig. `r fig_no`e shows the trend from `r fig_no`b again, now with imputed values. Fig. `r fig_no`e looks reasonable except for some disagreement prior to 1920, especially for 1918. The Baseball-Reference data (which is Cleveland only) is somewhat compromised by its definition of innings. A nine-inning game might be 27+27=54 outs, or 51 if the home team won. If ~50% of games are 3/54 shorter, duration per nine-innings is probably 4-5 minutes longer.

### `r fig_no`a. Coverage

```{r}
bind_rows(
  `Baseball-Reference` = bref_cle_season_ts %>% select(season, G, na_dur) %>% as_tibble(),
  `MLB API` = mlb_season_ts_0 %>% select(season, G) %>% as_tibble(),
  `Retrosheet (g)` = retro_season_g_ts %>% select(season, G, na_dur),
  `Retrosheet (e)` = retro_season_e_ts_0 %>% select(season, G, na_dur),
  .id = "Source"
) %>%
  mutate(
    Source = fct_inorder(Source),
    recorded = 1 - coalesce(na_dur, 0) / G
  ) %>%
  # replace_na(list(na_dur = 0)) %>%
  # mutate(recorded = 1 - na_dur / G) %>%
  ggplot(aes(x = season, y = recorded, fill = Source)) +
  geom_area(alpha = .80) +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  scale_y_continuous(labels = percent_format(1)) +
  scale_fill_manual(values = pal$mlb) +
  facet_wrap(facets = vars(Source)) +
  labs(
    title = "Game duration was not recorded regularly until after 1920.",
    subtitle = "Percent of games with no recorded duration.",
    x = NULL, y = NULL, fill = NULL
  ) +
  theme(axis.text.x = element_text(angle = 90, vjust = .5))
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`a. Coverage</b>. Baseball Reference and MLB API cover 1901-present; Retrosheet (g) covers 1900-2022; and Retrosheet (e) covers 1914-2022. Note Baseball-Reference is Cleveland games only.</font><br>

### `r fig_no`b. Trend

```{r warning=FALSE}
bind_rows(
  `Baseball-Reference (Cleve)` = bref_cle_season_ts %>% select(season, hrs_9I) %>% as_tibble(),
  `MLB API` = mlb_season_ts_0 %>% select(season, hrs_9I) %>% as_tibble(),
  `Retrosheet (g)` = retro_season_g_ts %>% select(season, hrs_9I),
  `Retrosheet (e)` = retro_season_e_ts_0 %>% select(season, hrs_9I),
  .id = "Source"
) %>%
  ggplot(aes(x = season, y = hrs_9I, color = Source, linewidth = Source, alpha = Source)) +
  geom_line() +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  scale_y_continuous(labels = pretty_hm) +
  scale_color_manual(values = pal$mlb) +
  scale_linewidth_manual(values = c(4, 1, 1, 1)) +
  scale_alpha_manual(values = c(.2, 1, 1, 1)) +
  labs(
    title = "MLB API has several dubious durations.",
    subtitle = "Four sources of avg duration per 9 innings of play.",
    x = NULL, y = NULL, color = NULL, linewidth = NULL, alpha = NULL
  ) +
  guides(fill = guide_legend(reverse = TRUE)) +
  theme(axis.text.x = element_text(angle = 90, vjust = .5))
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`b. Trend</b>. MLB API data has several problematic data points prior to 1930.</font><br>

### `r fig_no`c. Outliers

```{r}
dubious_seasons <- mlb_season_ts_0 %>% filter(hrs_9I < 1.5) %>% pull(season)

mlb_season_ts_0 %>%
  mutate(
    # dur_per_9 = time_length(dur / I * 9, "hours"),
    status = if_else(season %in% dubious_seasons, "Suspicious", "Probably good"),
    status = factor(status, levels = c("Suspicious", "Probably good"))
  ) %>%
  ggplot(aes(x = season, y = hrs_9I, color = status)) + 
  geom_point() +
  scale_color_manual(values = as.character(c(pal$oak["Y"], pal$cle["R"]))) +
  scale_y_continuous(labels = pretty_hm, limits = c(0, 3.5)) +
  labs(
    title = "Several MLB API game duration points are suspicious.",
    subtitle = "Hours per nine innings.",
    x = NULL, y = NULL, color = NULL
  ) +
  theme(legend.position = "top")
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`c. MLB API's Outliers</b>. A closer inspection identifies eight outliers.</font><br>

### `r fig_no`d. MICE

```{r warning=FALSE, echo=TRUE}
# Remove everything related to pitch counts prior to 1999.
x <- mlb_season_ts_0 %>%
  mutate(across(P_9I:P_pitcher, ~if_else(season < 1999, NA_real_, .)))

# Remove bad values from time columns
y <- x %>%
  mutate(across(hrs_9I:mins_PA, ~if_else(season %in% dubious_seasons, NA, .)))

# Use MICE to impute values in the time cols
z <- y %>%
  select(hrs_9I, mins_PA, PA_9I, pitchers_9I, PA_pitcher, season) %>%
  mice::mice(print = FALSE, seed = 123) %>%
  complete()

mlb_season_ts_1 <- mlb_season_ts_0
mlb_season_ts_1$hrs_9I <- z$hrs_9I
mlb_season_ts_1$mins_PA <- z$mins_PA
```

```{r}
mlb_season_ts_1 %>%
  mutate(
    status = if_else(season %in% dubious_seasons, "Imputed", "Probably good"),
    status = factor(status, levels = c("Imputed", "Probably good"))
  ) %>%
  ggplot(aes(x = season, y = hrs_9I, color = status)) + 
  geom_point() +
  scale_color_manual(values = as.character(c(pal$oak["Y"], pal$cle["R"]))) +
  scale_y_continuous(labels = pretty_hm, limits = c(0, 3.5)) +
  labs(
    title = "MLB API game durations with imputed values for outliers.", 
    x = NULL, y = NULL, color = NULL,
    subtitle = "Minutes per plate appearance"
  ) +
  theme(legend.position = "top")
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`d. MLB API with Imputations</b>. Using MICE to impute values produces realistic values.</font>

### `r fig_no`e. New Trend

```{r warning=FALSE}
bind_rows(
  `Baseball-Reference (Cleve)` = bref_cle_season_ts %>% select(season, hrs_9I) %>% as_tibble(),
  `MLB API` = mlb_season_ts_1 %>% select(season, hrs_9I) %>% as_tibble(),
  `Retrosheet (g)` = retro_season_g_ts %>% select(season, hrs_9I),
  `Retrosheet (e)` = retro_season_e_ts_0 %>% select(season, hrs_9I),
  .id = "Source"
) %>%
  ggplot(aes(x = season, y = hrs_9I, color = Source, linewidth = Source, alpha = Source)) +
  geom_line() +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  scale_y_continuous(labels = pretty_hm) +
  scale_color_manual(values = pal$mlb) +
  scale_linewidth_manual(values = c(4, 1, 1, 1)) +
  scale_alpha_manual(values = c(.2, 1, 1, 1)) +
  labs(
    title = "Sources are in near perfect alignment.",
    subtitle = "Avg duration per 9 innings of play, MLB API imputed.",
    x = NULL, y = NULL, color = NULL, linewidth = NULL, alpha = NULL
  ) +
  guides(fill = guide_legend(reverse = TRUE)) +
  theme(axis.text.x = element_text(angle = 90, vjust = .5))
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`e. New Trend</b>. The imputed data set now fits nicely with the others.</font>

## Plate Appearances {.tabset .tabset-fade .tabset-pills}

```{r}
fig_no <- fig_no + 1
```

My Baseball-Reference data under-counts plate appearances (PAs). The [MLB definition of PA](https://www.mlb.com/glossary/standard-stats/plate-appearance) is runs + outs + LOB. I defined PA as AB + sacs + BBs because I didn't have LOB. Fig. `r fig_no`a shows the trend of PA per nine-innings. MLB API is unrealistic prior to 1910. It is slightly lower than the two Retrosheet sources for the other years. Retrosheet looks like the more reliable source, but the trough between 1901 and 1920 is strange. Why did PA per 9 fall and rise like that? I will impute values for MLB API for 1901-1911 using Retrosheet (g) (Fig. `r fig_no`b), but it seems like only 1920+ data is trustworthy.

What does Fig. `r fig_no`b tell us? PA has a long gradual decline from 1935-1968 totaling nearly six less batters. If each batter takes around 2.5 minutes, that might reduce game times by 15 minutes. There is a partial recovery of five batters by 2000, then another decline. Changes in PA are an unlikely factor game duration.

### `r fig_no`a. Trend

```{r warning=FALSE}
bind_rows(
  `Baseball-Reference (Cleve)` = bref_cle_season_ts %>% select(season, PA_9I) %>% as_tibble(),
  `MLB API` = mlb_season_ts_1 %>% select(season, PA_9I) %>% as_tibble(),
  `Retrosheet (g)` = retro_season_g_ts %>% select(season, PA_9I) %>% as_tibble(),
  `Retrosheet (e)` = retro_season_e_ts_0 %>% select(season, PA_9I) %>% as_tibble(),
  .id = "Source"
) %>%
  # pivot_wider(values_from = PA_9I, names_from = Source)
  ggplot(aes(x = season, y = PA_9I, color = Source, linewidth = Source, alpha = Source)) +
  geom_line() +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  scale_y_continuous() +
  scale_color_manual(values = pal$mlb) +
  scale_linewidth_manual(values = c(4, 1, 1, 1)) +
  scale_alpha_manual(values = c(.2, 1, 1, 1)) +
  labs(
    title = "PA problematic before 1920.",
    subtitle = "Avg plate appearances per 9 innings of play.",
    x = NULL, y = NULL, color = NULL, linewidth = NULL, alpha = NULL
  ) +
  guides(fill = guide_legend(reverse = TRUE)) +
  theme(axis.text.x = element_text(angle = 90, vjust = .5))
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`a. Trend</b>. MLB API plate appearances per nine-innings diverges from other sources prior to 1910, but all values seem strange prior to 1920. Note Baseball-Reference is Cleveland games only, and uses rough approximation of PA and innings.</font><br>

### `r fig_no`b. New Trend

```{r echo=TRUE}
mlb_season_ts_2 <- mlb_season_ts_1 %>%
  left_join(retro_season_g_ts %>% select(season, retro_PA_9I = PA_9I), 
             by = join_by(season)) %>%
  mutate(PA_9I = if_else(season %in% 1901:1911, retro_PA_9I, PA_9I)) %>%
  select(-retro_PA_9I)
```

```{r warning=FALSE}
bind_rows(
  `MLB API` = mlb_season_ts_2 %>% select(season, PA_9I) %>% as_tibble(),
  `Retrosheet (g)` = retro_season_g_ts %>% select(season, PA_9I) %>% as_tibble(),
  `Retrosheet (e)` = retro_season_e_ts_0 %>% select(season, PA_9I) %>% as_tibble(),
  .id = "Source"
) %>%
  ggplot(aes(x = season, y = PA_9I, color = Source)) +
  geom_line(linewidth = 1) +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  scale_y_continuous() +
  scale_color_manual(values = pal$mlb[2:4]) +
  labs(
    title = "Sources are in close alignment.",
    subtitle = "Avg plate appearances per 9 innings of play, MLB API imputed.",
    x = NULL, y = NULL, color = NULL, linewidth = NULL, alpha = NULL
  ) +
  guides(fill = guide_legend(reverse = TRUE)) +
  theme(axis.text.x = element_text(angle = 90, vjust = .5))
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`b. Trend after Imputation</b>. MLB API data imputed for 1901-1911. Baseball-Reference not shown.</font><br>

## Pitches {.tabset .tabset-fade .tabset-pills}

```{r}
fig_no <- fig_no + 1
```

There is no recorded pitch counts until 1950 and until 1988 the values are less than half of what seems reasonable (Fig. `r fig_no`a). The 1988-1998 values are still low and variable. Only 1999-present are consistent. I will null out all pitch counts prior to 1999. The pared down data set shows a small upward trend of about 10 additional pitches per nine innings since 1999 (Fig. `r fig_no`b). The MLB-API data is lower by about 8 pitches per nine-innings. The lower pitch counts for MLB API along with the lower PAs make me suspect MLB API does not account for games when the home team does not bat in the bottom of the inning.

### `r fig_no`a. Trend

```{r warning=FALSE}
bind_rows(
  `MLB API` = mlb_season_ts_2 %>% select(season, P_9I) %>% as_tibble(),
  `Retrosheet (e)` = retro_season_e_ts_0 %>% select(season, P_9I) %>% as_tibble(),
  .id = "Source"
) %>%
  mutate(
    status = factor(
      case_when(season < 1988 ~ "Definitely bad",
                season < 1999 ~ "Suspicious",
                TRUE ~ "Probably good"),
      levels = c("Definitely bad", "Suspicious", "Probably good")
    )
  ) %>%
  ggplot(aes(x = season, y = P_9I, shape = status, color = Source)) + 
  geom_point() +
  scale_color_manual(values = pal$mlb[2:3]) +
  labs(
    title = "Pitch count data is not reliable until 1999.", 
    subtitle = "Average pitches per nine innings.",
    x = NULL, y = NULL, color = NULL, shape = NULL
  )
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`a. Trend</b>. Pitch count per nine-innings. Shapes indicate plausibility.</font><br>

### `r fig_no`b. New Trend

```{r}
mlb_season_ts_3 <- mlb_season_ts_2 %>% 
  mutate(
    across(c(P_9I, P_PA, P_pitcher, secs_P), ~if_else(season < 1999, NA_real_, .))
  )

retro_season_e_ts <- retro_season_e_ts_0 %>%
  mutate(
    across(c(P_9I, P_PA, P_pitcher, secs_P), ~if_else(season < 1999, NA_real_, .))
  )
```

```{r warning=FALSE}
bind_rows(
  `MLB API` = mlb_season_ts_3 %>% select(season, P_9I) %>% as_tibble(),
  `Retrosheet (e)` = retro_season_e_ts %>% select(season, P_9I) %>% as_tibble(),
  .id = "Source"
) %>%
  ggplot(aes(x = season, y = P_9I, color = Source)) + 
  geom_point() +
  geom_smooth(method = "loess", formula = "y~x", se = TRUE) +
  scale_color_manual(values = pal$mlb[2:3]) +
  labs(
    title = "Sources are in close alignment.",
    subtitle = "Avg pitches per nine innings, MLB API and Retrosheet (e) pared.",
    x = NULL, y = NULL, color = NULL
  )
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`b. New Trend</b>. Data sets after removing implausible data. MLB API is still lower than Retrosheet by about 8 pitches per game.</font><br>

## Pitchers {.tabset .tabset-fade .tabset-pills}

```{r}
fig_no <- fig_no + 1
```

The three sources of pitcher counts are consistent. MLB API and Retrosheet (g) start in 1901, and Retrosheet (e) starts in 1914 (Fig. `r fig_no`a **Total**). Retrosheet (e) has details about when pitching substitutions take place. Fig. `r fig_no`b breaks out mid-inning relief appearances. The drop-off of begin-inning substitutions in 1973 coincides with the introduction of the designated hitter. The temporary increase in mid-inning substitutions in 2020 might have something to do with COVID or the three-batter rule (discussed later). 

### `r fig_no`a. Total

```{r warning=FALSE}
bind_rows(
  `MLB API` = mlb_season_ts_3 %>% select(season, pitchers_9I) %>% as_tibble(),
  `Retrosheet (g)` = retro_season_g_ts %>% select(season, pitchers_9I) %>% as_tibble(),
  `Retrosheet (e)` = retro_season_e_ts %>% select(season, pitchers_9I) %>% as_tibble(),
  .id = "Source"
) %>%
  ggplot(aes(x = season, y = pitchers_9I, color = Source)) + 
  geom_point() +
  scale_y_continuous(labels = comma_format(1), breaks = 0:10) +
  coord_cartesian(ylim = c(0, 10)) +
  scale_color_manual(values = pal$mlb[2:4]) +
  labs(
    title = "Count of pitchers is consistent across time and sources.",
    subtitle = "Avg pitchers per nine innings.",
    x = NULL, y = NULL, color = NULL
  )
```

<font size=-1 color="#666666"><b>Fig. `r (fig_no <- fig_no + 1)`a. Total</b>. All three sources have pitchers per game. MLB API stretches back to 1900 and includes the current season YTD. Retrosheet (g) starts in 1901. Retrosheet (e) starts in 1914.</font>

### `r fig_no`b. Relievers

```{r warning=FALSE}
retro_season_e_ts %>%
  mutate(Total = pitchers_9I - 2) %>%
  rename(`Mid-inning` = pitsubsmi_9I, `Start-inning` = pitsubsbi_9I) %>%
  pivot_longer(cols = c(Total, `Mid-inning`, `Start-inning`), names_to = "Source", values_to = "y") %>%
  ggplot(aes(x = season, y = y, color = Source)) + 
  geom_point() +
  scale_y_continuous(labels = comma_format(1), breaks = 0:10) +
  coord_cartesian(ylim = c(0, NA)) +
  scale_color_manual(values = pal$mlb) +
  labs(
    title = "Increased usage of relievers to start an inning.",
    subtitle = "Relief pitcher appearances per nine innings.",
    x = NULL, y = NULL, color = NULL
  )
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`b. Relievers</b>. Only Retrosheet (e) provides this level of detail.</font>

## Stolen Bases {.tabset .tabset-fade .tabset-pills}

```{r}
fig_no <- fig_no + 1
```

The stolen bases are less important because they shouldn't affect game duration, but changes to its frequency are a bi-product of the 2023 rule changes related to base width and pick-off attempts. The caught-steeling data prior to 1926 is clearly not to be trusted, but what about the caught-stealing data from 1926 - 1950 (Fig. `r fig_no`a)? I'm inclined to null out all of the caught-steeling data prior to 1951. There are also two outliers in 1914 and 1915. I don't have an explanation for why they would pop up like that and it's a big change from the prior trend. The 1914 outlier looks like it could be stolen base _attempts_. I will impute values for 1914 and 1915 (Fig. `r fig_no`b).

### `r fig_no`a. Trend 

```{r warning=FALSE}
mlb_season_ts_3 %>% 
  rename(SB = SB_9I, CS = CS_9I) %>%
  pivot_longer(cols = c(SB, CS)) %>%
  ggplot(aes(x = season, y = value, color = name)) +
  geom_point() +
  scale_color_manual(values = pal$mlb) +
  coord_cartesian(ylim = c(0, 4)) +
  labs(
    title = "Stolen base data is suspicious prior to 1951.",
    subtitle = "Stolen bases per nine innings.",
    x = NULL, y = NULL, color = NULL
  )
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`a. Stolen bases per nine innings</b>.</font>

### `r fig_no`b. Redacted and Imputed

```{r echo=TRUE}
# Remove everything related to pitch counts prior to 1999.
mlb_season_ts <- mlb_season_ts_3 %>%
  mutate(
    SB_9I = if_else(season %in% 1914:1915, NA_real_, SB_9I),
    CS_9I = if_else(season < 1951, NA_real_, CS_9I)
  )

# Use MICE to impute values in the time cols
x <- mlb_season_ts %>%
  select(season, SB_9I) %>%
  mice::mice(print = FALSE, seed = 123) %>%
  complete()

mlb_season_ts$SB_9I <- x$SB_9I
```

```{r warning=FALSE}
mlb_season_ts %>% 
  rename(SB = SB_9I, CS = CS_9I) %>%
  pivot_longer(cols = c(SB, CS)) %>%
  ggplot(aes(x = season, y = value, color = name)) +
  geom_point() +
  scale_color_manual(values = pal$mlb) +
  coord_cartesian(ylim = c(0, 4)) +
  labs(
    title = "Stolen base data, 1914-15 imputed, CS redacted prior to 1951.",
    subtitle = "Stolen bases per nine innings.",
    x = NULL, y = NULL, color = NULL
  )
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`b. Modified</b>. Stolen bases per nine innings, after removing caught-steeling prior to 1951 and imputing successes in 1914 and 1915.</font>

# Save Data

Next I'll dive into features, but there are no more modifications to the underlying data. Save it back to disk.

```{r echo=TRUE, class.source='fold-show'}
saveRDS(bref_cle_season_ts, file.path("../data", "bref_cle_season_ts.rds"))
saveRDS(retro_season_e_ts, file.path("../data", "retro_season_e_ts.rds"))
saveRDS(retro_season_g_ts, file.path("../data", "retro_season_g_ts.rds"))
saveRDS(mlb_season_ts, file.path("../data", "mlb_season_ts.rds"))
```

# Feature Exploration

These subsections take a closer look at the time-related features in the data set.

## Game Duration

```{r}
fig_no <- fig_no + 1

max_avg <- mlb_season_ts %>% slice_max(order_by = hrs_9I)
avg_2023 <- mlb_season_ts %>% filter(season == 2023)
max_m_23 <- (max_avg[1, ]$hrs_9I - avg_2023[1, ]$hrs_9I) * 60
```

Games typically lasted around two hours up to 1945. Then over the course of ten seasons game times shot up to 2.5 hours. In the late 1970s games began another decade-long increase, hitting `r mlb_season_ts %>% filter(season == 1988) %>% pull(hrs_9I) %>% pretty_hm(verbose = TRUE)` in 1988. After 2011, game times began a third period of increase, reaching `r max_avg %>% pull(hrs_9I) %>% pretty_hm(verbose = TRUE)` in `r max_avg %>% pull(season)`. Several rule changes introduced in 2023 appear to have halted the advance. Average game times in 2023 through `r max(bref_cle_games$game_date) %>% format("%b %d")` have been `r avg_2023 %>% pull(hrs_9I) %>% pretty_hm(verbose = TRUE)`, a reduction of `r comma(max_m_23)` minutes. 

```{r}
knot_yrs <- c(1901, 1944, 1955, 1978, 1988, 2011, 2021, 2023)

model(mlb_season_ts, TSLM(hrs_9I ~ trend(knots = knot_yrs))) %>%
  augment() %>%
  ggplot(aes(x = season)) +
  geom_line(aes(y = .fitted), color = pal$sfg["O"], linewidth = 1.5, alpha = .5) +
  geom_point(aes(y = hrs_9I), color = pal$sfg["B"]) +
  geom_text(data = mlb_season_ts %>% filter(season %in% knot_yrs), 
            aes(y = hrs_9I + if_else(season == 2021, +.15, -.15), label = pretty_hm(hrs_9I)), 
            family = "Rockwell", size = 3.5, color = pal$sfg["B"], hjust = 0) + 
  geom_point(data = mlb_season_ts %>% filter(season %in% knot_yrs), 
             aes(y = hrs_9I),
             color = pal$sfg["O"], size = 2) +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  scale_y_continuous(limits = c(0, NA), n.breaks = 8, labels = pretty_hm) +
  coord_cartesian(xlim = c(1900, 2030), ylim = c(0, 3.5)) +
  labs(
    title = "Pace of play slowed in three phases: 1945-55, 1978-88, and 2011-21.",
    subtitle = "Average game duration per nine innings, 1901-2023.",
    x = NULL, y = NULL
  )
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`. Average elapsed time per nine innings of play</b>. Source: MLB-API, 1901-`r max(bref_cle_games$game_date) %>% format("%b %d, %Y")`.</font>

Doug Pappas chronicled the increased playing times in the Summer 1995 issue of Outside the Lines, the SABR Business of Baseball Committee newsletter. Pappas found a Sporting News article from 1962 blaming longer game times on full counts, pitching changes, and delays between pitches.^[Doug Pappas. Originally published in the Summer 1995 issue of Outside the Lines, the SABR Business of Baseball Committee newsletter. [html](http://roadsidephotos.sabr.org/baseball/95gamelength.htm).] I'll consider these hypotheses and others below.

## Batting {.tabset .tabset-fade .tabset-pills}

```{r include=FALSE}
fig_no <- fig_no + 1

minyr <- 1920
maxyr <- max(retro_season_e_ts$season)

max_K <- retro_season_e_ts %>% filter(SO_9I == max(retro_season_e_ts$SO_9I))

# Fit a univariate OLS and pull the R^2. R^2 is .81! I would use fable::TSLM(),
# but it does not support Cook's distance.
lm_K <- retro_season_e_ts %>% 
  filter(season >= minyr) %>%
  lm(hrs_9I ~ log(SO_9I), data = .)
r2_K <- glance(lm_K) %>% pull(adj.r.squared)
# The plot has evidence of autocorrelation - there are other factors impacting
# the trend. Including a t-1 term increases R^2 to .95. The standard error of 
# the regression is .10 - about 6 minutes.
lm_K_trend <- retro_season_e_ts %>% 
  filter(season >= minyr) %>%
  model(TSLM(hrs_9I ~ log(SO_9I) + trend()))
lm_K_trend %>% report()

# For plots
lbl_yr <- c(minyr, maxyr)
retro_minyr <- retro_season_e_ts %>% filter(season == minyr)

plot_trend_vs_dur <- function(dat) {
  dat %>%
    filter(season >= minyr) %>%
    mutate(std_hrs_9I = hrs_9I / retro_minyr$hrs_9I) %>%
    ggplot(aes(x = season)) +
    geom_line(aes(y = std_hrs_9I), color = pal$sfg["B"], linewidth = 2, alpha = 0.15) +
    geom_line(aes(y = std_y), color = pal$sfg["O"], linewidth = 1) +
    geom_point(aes(y = if_else(season %in% lbl_yr, std_hrs_9I, NA_real_)),
               color = pal$sfg["B"], size = 2) +
    geom_point(aes(y = if_else(season %in% lbl_yr, std_y, NA_real_)),
               color = pal$sfg["O"], size = 2) +
    geom_text(aes(x = season + 1,
                  y = if_else(season %in% lbl_yr, std_y + .05, NA_real_),
                  label = comma(act_y, .1)),
              family = "Rockwell", size = 3.5, color = pal$sfg["O"], hjust = 0) +
    geom_text(aes(x = season - 3,
                  y = if_else(season %in% lbl_yr, std_hrs_9I - .05, NA_real_),
                  label = pretty_hm(hrs_9I, verbose = FALSE)),
              family = "Rockwell", size = 3.5, color = pal$sfg["B"], hjust = 0) +
    scale_x_continuous(breaks = seq(minyr, 2030, 10)) +
    scale_y_continuous(labels = ~ percent(.x - 1)) +
    coord_cartesian(xlim = c(minyr, 2030)) +
    labs(x = NULL, y = NULL)
}
```

Plate appearances (PA) per nine innings of play have changed little over time (Fig. `r fig_no`a). Walks increased with game duration, at least through the 1940s, suggesting strategy was becoming increasingly important (Fig. `r fig_no`b). Strikeouts in particular have increased with game times (Fig. `r fig_no`c). Of the main offense metrics, strikeouts come closest to matching the trend in game duration. Strikeouts alone can explain about `r percent(r2_K, 1)` of the variation in average game duration (Fig. `r fig_no`d). Finally, hits (H) have fluctuated around 17 per nine innings since the 1940s (Fig. `r fig_no`e).

### `r fig_no`a. PA

```{r warning=FALSE}
retro_season_e_ts %>%
  mutate(act_y = PA_9I, std_y = PA_9I / retro_minyr$PA_9I) %>%
  plot_trend_vs_dur() +
  labs(title = glue("Plate appearances and game duration, {minyr}-{maxyr}."), 
       subtitle = "Percent change")
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`a. Plate appearances (orange) and game duration (gray), per nine-innings</b>. Plate appearances are nearly unchanged.<br> Source: Retrosheet Game.</font>

### `r fig_no`b. BB

```{r warning=FALSE}
retro_season_g_ts %>%
  mutate(act_y = BB_9I, std_y = BB_9I / retro_minyr$BB_9I) %>%
  plot_trend_vs_dur() +
  labs(title = glue("Walks and game duration, {minyr}-{maxyr}."), 
       subtitle = "Percent change")
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`b. Walks (orange) and game duration (gray), per nine-innings</b>. Walks increased with game duration until 1950.<br> Source: Retrosheet Game.</font>

### `r fig_no`c. SO

```{r warning=FALSE}
retro_season_g_ts %>%
  mutate(act_y = SO_9I, std_y = SO_9I / retro_minyr$SO_9I) %>%
  plot_trend_vs_dur() +
  labs(title = glue("Strikeouts and game duration, {minyr}-{maxyr}."), 
       subtitle = "Percent change")
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`c. Strikeouts (orange) and game duration (gray), per nine-innings</b>. Strikeouts climbed gradually after the dead-ball era, then more sharply after WWII. They declined from the mid-60's to 1980, but have climbed since then, peaking at `r comma(max_K[1, ]$SO_9I, .1)` in `r max_K[1, ]$season`.<br> Source: Retrosheet Game.</font>

### `r fig_no`d. SO vs Hrs

```{r warning=FALSE}
bind_cols(
  retro_season_e_ts %>% filter(season >= minyr) %>% select(season, SO_9I), 
  broom::augment(lm_K)
) %>%
  mutate(influential = .cooksd > 4 * mean(.cooksd)) %>%
  ggplot(aes(x = log(SO_9I), y = hrs_9I)) +
  geom_point(aes(color = influential), show.legend = FALSE) +
  geom_text_repel(aes(label = if_else(influential, season, NA_real_)), 
                  size = 3, color = pal$mlb[5]) +
  geom_smooth(method = "lm", formula = "y ~ x", color = pal$oak["Y"]) +
  scale_color_manual(values = pal$mlb[c(4, 5)]) +
  labs(
    title = "Strikeouts are correlated with longer game times.",
    subtitle = glue("Adjusted R2 = {comma(glance(lm_K)[1, ]$adj.r.squared, .01)}."),
    x = "K's per 9 (log)", y = "hours per 9"
  )
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`d. Log of strikeouts per nine innings vs game duration</b>. Blue labeled points are influential. Fit line shows evidence of autocorrelation, suggesting there are other factors influencing game times. R-squared is `r comma(r2_K, .01)`. 
<br>Source: MLB Stats API.</font>

### `r fig_no`e. H

```{r warning=FALSE}
retro_season_g_ts %>%
  mutate(act_y = H_9I, std_y = H_9I / retro_minyr$H_9I) %>%
  plot_trend_vs_dur() +
  labs(title = glue("Hits and game duration, {minyr}-{maxyr}."), 
       subtitle = "Percent change")
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`e. Hits (orange) and game duration (gray), per nine-innings</b>. Hits decreased to about 17 by 1942 and have varied little since then.<br> Source: Retrosheet Game.</font>

## Pitching {.tabset .tabset-fade .tabset-pills}

```{r include=FALSE}
fig_no <- fig_no + 1

# Fit a univariate OLS and pull the R^2. R^2 is .86!
lm_pitchers <- retro_season_e_ts %>% 
  filter(season >= minyr) %>% 
  lm(hrs_9I ~ pitchers_9I, data = .)
r2_pitchers <- glance(lm_pitchers) %>% pull(adj.r.squared)

# That was promising. What if we combine pitchers_9I with SO_9I from the
# batting exploration?
lm_pK <- retro_season_e_ts %>% 
  filter(season >= minyr) %>% 
  lm(hrs_9I ~ log(SO_9I) + pitchers_9I, data = .)
r2_pK <- glance(lm_pK) %>% pull(adj.r.squared)

# Now let's model mid-inning relievers.
lm_pitsubsmi <- retro_season_e_ts %>%
  filter(season >= minyr) %>%
  lm(hrs_9I ~ pitsubsmi_9I, data = .)
r2_pitsubsmi <- glance(lm_pitsubsmi) %>% pull(adj.r.squared)
```
The increase in pitcher appearances tracked game duration fairly well until they diverged around 1990 (Fig. `r fig_no`a). The association is fairly linear, although there is evidence of autocorrelation (Fig. `r fig_no`b). Fitting a model of game duration vs pitching appearances _and_ strikeouts improves the explanatory power (R2 = `r percent(r2_pK, 1)`), but autocorrelation is still evident (Fig. `r fig_no`c). I don't know if it is the physical act of replacing a pitcher, or just the strategic activity implied by it (mound visits by coach and/or catcher, taking time between pitches, etc.). Fig. `r fig_no`d focuses on the physical act by showing just the mid-inning pitching changes. Expressed as a percentage change, mid-inning relievers is choppy. The R2 on a model with just this variable is `r percent(r2_pitsubsmi, 1)`. Fig. `r fig_no`e shows the relationship with pitch counts, but the data set is limited, and the trend is flat.

### `r fig_no`a. Pitchers

```{r warning=FALSE}
retro_season_e_ts %>%
  mutate(act_y = pitchers_9I, std_y = pitchers_9I / retro_minyr$pitchers_9I) %>%
  plot_trend_vs_dur() +
  labs(title = glue("Pitcher appearances and game duration, {minyr}-{maxyr}."), 
       subtitle = "Percent change")
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`a. Pitcher appearances (orange) and game duration (gray), per nine-innings</b>. The trends diverge around 1990.<br> Source: Retrosheet Game.</font>

### `r fig_no`b. Pitches vs Hrs

```{r warning=FALSE}
bind_cols(
  retro_season_e_ts %>% filter(season >= minyr) %>% select(season), 
  augment(lm_pitchers)
) %>%
  mutate(influential = .cooksd > 4 * mean(.cooksd)) %>%
  ggplot(aes(x = pitchers_9I, y = hrs_9I)) +
  geom_point(aes(color = influential), show.legend = FALSE) +
  geom_text_repel(aes(label = if_else(influential, season, NA_real_)), 
                  size = 3, color = pal$mlb[5]) +
  geom_smooth(method = "lm", formula = "y ~ x", color = pal$oak["Y"]) +
  scale_color_manual(values = pal$mlb[c(4, 5)]) +
  labs(
    title = "Pitching appearances are correlated with longer game times.",
    subtitle = glue("Adjusted R2 = {comma(glance(lm_pitchers)[1, ]$adj.r.squared, .01)}."),
    x = "Pitchers per 9", y = "hours per 9"
  )
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`b. Log of pitchers per nine innings vs game duration</b>. Blue labeled points are influential. Fit line shows evidence of autocorrelation, suggesting there are other factors influencing game times. R-squared is `r comma(r2_pitchers, .01)`. 
<br>Source: Retrosheet Game.</font>

### `r fig_no`c. Pitchers + K

```{r}
bind_cols(
  retro_season_e_ts %>% filter(season >= minyr) %>% select(season), 
  broom::augment(lm_pK)
) %>%
  ggplot(aes(x = season)) +
  geom_point(aes(y = hrs_9I)) +
  geom_line(aes(y = .fitted), color = pal$oak["Y"], linewidth = 1) +
  labs(
    title = "Fitted values: pitchers & strikeouts.",
    subtitle = glue("Adjusted R2 = {comma(glance(lm_pK)[1, ]$adj.r.squared, .01)}."),
    x = "season", y = "hours per 9"
  )
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`c. Fitted values of model of hrs-per-9 vs log of strikeouts plus pitcher appearances.</b> Model still shows signs of autocorrelation. R-squared is `r comma(r2_pK, .01)`. 
<br>Source: Retrosheet Event.</font>

### `r fig_no`d. Mid-inning

```{r warning=FALSE}
retro_season_e_ts %>% 
  mutate(act_y = pitsubsmi_9I, std_y = pitsubsmi_9I / retro_minyr$pitsubsmi_9I) %>%
  plot_trend_vs_dur() +
  labs(title = glue("Mid-inning relief appearances and game duration, {minyr}-{maxyr}."), 
       subtitle = "Percent change")
```

<font size=-1 color="#666666"><b>Fig. `r fig_no`d. Mid-inning relief appearances (orange) and game duration (gray), per nine-innings</b>.<br> Source: Retrosheet Event.</font>

### `r fig_no`e. Pitches

```{r warning=FALSE}
P_9I_1999 <- retro_season_e_ts %>% filter(season == 1999) %>% pull(P_9I)
retro_season_e_ts %>% 
  mutate(act_y = P_9I, std_y = P_9I / P_9I_1999) %>%
  plot_trend_vs_dur() +
  labs(title = glue("Pitch count and game duration, {minyr}-{maxyr}."), 
       subtitle = "Percent change")
```
<font size=-1 color="#666666"><b>Fig. `r fig_no`e. Pitch counts (orange) and game duration (gray), per nine-innings</b>.<br> Source: Retrosheet Event.</font>

# Appendix A: Correlation Matrix

```{r}
mlb_season_ts %>% 
  select(ends_with("_9I")) %>%
  cor() %>%
  corrplot::corrplot(
    method = "square",
    type = "lower",
    diag = FALSE
  )
```

# Appendix B: Benchmark Forecasts

The following sections will use time series analysis to forecast attendance. Some forecasting methods are extremely simple and surprisingly effective. I'll keep them in mind as benchmarks to evaluate the more sophisticated methods. I denote a forecast value with the "hat" notation, $\hat{y}_{T+h|T}$. The subscript $T + h$ means forecasting $h$ periods beyond the observed time series $T$. $\hat{y}_{T+h|T}$ means an $h$ period forecast beyond $T$ based on data through period $T$.

* The **average method** projects the historical average, $\hat{y}_{T+h|T} = \bar{y}.$ 
* The **naive method** projects the last observation, $\hat{y}_{T+h|T} = y_T.$ 
* The **seasonal naive method** projects the last seasonal observation, $\hat{y}_{T+h|T} = y_{T+h-m(k+1)}$. 

The methods can include drift, a straight line from the first and last observation, $\hat{y}_{T+h|T} = y_T + h\left(\frac{y_T - y_1}{T-1}\right).$

```{r warning=FALSE}
bref_cle_season_ts %>%
  # fill_gaps() %>%
  model(Average = MEAN(hrs_9I),
        Naive = NAIVE(hrs_9I),
        RW = RW(hrs_9I ~ drift())) %>%
  forecast(h = 10) %>%
  ggplot(aes(x = season)) +
  geom_line(aes(y = .mean, color = .model)) +
  geom_line(data = bref_cle_season_ts %>% filter(season >= 1911) %>% fill_gaps(), aes(y = hrs_9I)) +
  scale_y_continuous(limits = c(0, NA)) +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  labs(title = "Three benchmark Tribe game duration forecasts.",
       subtitle = "Average, naive, and random walk forecasts, h = 10 seasons.", 
       x = NULL, y = "Hours", color = NULL,
       caption = "source: www.baseball-reference.com.") +
  guides(color = guide_legend(title = "Forecast"))
```

# Appendix C: Stolen Bases

```{r warning=FALSE}
interesting <- c(1987, 2023)

my_dat <- mlb_season_ts %>% filter(season >= 1951) %>% mutate(attempts = SB_9I + CS_9I)
my_cols <- my_dat %>% pivot_longer(cols = c(SB_9I, CS_9I))
my_line <- my_dat %>% mutate(r = SB_9I / (SB_9I + CS_9I))
my_mult <- 2.5
min_r <- my_line %>% filter(r == min(my_line$r)) %>% 
  mutate(my_y = r * my_mult, my_lbl_y = my_y - 0.15, my_hjust = 0, my_lbl = glue("{percent(r, 1)}"))
max_r <- my_line %>% filter(r == max(my_line$r)) %>% 
  mutate(my_y = r * my_mult, my_lbl_y = my_y + 0.15, my_hjust = 1, my_lbl = glue("{percent(r, 1)} success"))
min_s <- my_dat %>% filter(attempts == min(my_dat$attempts)) %>% 
  mutate(my_lbl_x = season, my_lbl_y = attempts - 0.15, my_hjust = 0, 
         my_lbl = glue("{comma(attempts, .1)} attempts per 9"))
max_s <- my_dat %>% filter(attempts == max(my_dat$attempts)) %>% 
  mutate(my_lbl_x = season + 1, my_lbl_y = attempts + 0.05, my_hjust = 0, 
         my_lbl = glue("{comma(attempts, .1)} ({season})"))
sb_23 <- my_dat %>% filter(season == 2023) %>%
  mutate(my_lbl_x = season - .5, my_lbl_y = attempts - 0.15, my_hjust = 1, 
         my_lbl = glue("{comma(attempts, .1)} attempts"))

my_cols %>% ggplot(aes(x = season)) +
  geom_area(aes(y = value, fill = name), color="gray80", alpha = .85) +
  geom_line(data = my_line, aes(y = r * my_mult), show.legend = FALSE, 
            linewidth = 1, color = pal$oak["Y"]) +
  geom_point(data = bind_rows(min_r, max_r), aes(y = my_y), size = 2, color = pal$oak["Y"]) +
  geom_text(
    data = bind_rows(min_r, max_r),
    aes(y = my_lbl_y, label = my_lbl, hjust = my_hjust),
    size = 4.5, family = "Rockwell Condensed", color = "darkgoldenrod"
    ) +
  geom_point(data = bind_rows(min_s, max_s, sb_23), aes(y = attempts),
             color = pal$old["grass"], size = 2) +
  geom_text(
    data = bind_rows(min_s, max_s, sb_23),
    aes(x = my_lbl_x, y = my_lbl_y, label = my_lbl, hjust = my_hjust),
    size = 4.5, family = "Rockwell Condensed", color = pal$old["grass"]
    ) +
  scale_y_continuous(
    limits = c(0, my_mult), labels = comma_format(.1), breaks = seq(0, 2.5, .5)
  ) +
  scale_x_continuous(breaks = seq(1950, 2030, 10)) +
  scale_fill_manual(
    values = as.character(pal$old[c("plate", "dust")]),
    labels = c("caught steeling", "swiped")
  ) +
  labs(
    title = str_wrap("Stolen bases way up thanks to rule changes this year.", 80),
    subtitle = glue("Stolen bases per nine innings, 1951 - {format(last_run, '%b %d, %Y')}."),
    caption = "Source: MLB Stats API.",
    x = NULL, y = NULL, fill = NULL
  )
```

```{r social-media, warning=FALSE, fig.width=3, fig.height=5, eval=FALSE}
interesting <- c(1987, 2023)

my_dat <- mlb_season_ts %>% filter(season >= 1951) %>% mutate(attempts = SB_9I + CS_9I)
my_cols <- my_dat %>% pivot_longer(cols = c(SB_9I, CS_9I))
my_line <- my_dat %>% mutate(r = SB_9I / (SB_9I + CS_9I))
my_mult <- 2.5
min_r <- my_line %>% filter(r == min(my_line$r)) %>% 
  mutate(my_y = r * my_mult, my_lbl_y = my_y - 0.15, my_hjust = 0, my_lbl = glue("{percent(r, 1)}"))
max_r <- my_line %>% filter(r == max(my_line$r)) %>% 
  mutate(my_y = r * my_mult, my_lbl_y = my_y + 0.15, my_hjust = 1, my_lbl = glue("{percent(r, 1)} success"))
min_s <- my_dat %>% filter(attempts == min(my_dat$attempts)) %>% 
  mutate(my_lbl_x = season, my_lbl_y = attempts - 0.15, my_hjust = 0, 
         my_lbl = glue("{comma(attempts, .1)} attempts per 9"))
max_s <- my_dat %>% filter(attempts == max(my_dat$attempts)) %>% 
  mutate(my_lbl_x = season + 1, my_lbl_y = attempts + 0.05, my_hjust = 0, 
         my_lbl = glue("{comma(attempts, .1)} ({season})"))
sb_23 <- my_dat %>% filter(season == 2023) %>%
  mutate(my_lbl_x = season - .5, my_lbl_y = attempts - 0.15, my_hjust = 1, 
         my_lbl = glue("{comma(attempts, .1)} attempts"))

my_cols %>% ggplot(aes(x = season)) +
  geom_area(aes(y = value, fill = name), color="gray80", alpha = .85) +
  geom_line(data = my_line, aes(y = r * my_mult), show.legend = FALSE, 
            linewidth = 1, color = pal$oak["Y"]) +
  geom_point(data = bind_rows(min_r, max_r), aes(y = my_y), size = 2, color = pal$oak["Y"]) +
  geom_text(
    data = bind_rows(min_r, max_r),
    aes(y = my_lbl_y, label = my_lbl, hjust = my_hjust),
    size = 4.5, family = "Rockwell Condensed", color = "darkgoldenrod"
    ) +
  geom_point(data = bind_rows(min_s, max_s, sb_23), aes(y = attempts),
             color = pal$old["grass"], size = 2) +
  geom_text(
    data = bind_rows(min_s, max_s, sb_23),
    aes(x = my_lbl_x, y = my_lbl_y, label = my_lbl, hjust = my_hjust),
    size = 4.5, family = "Rockwell Condensed", color = pal$old["grass"]
    ) +
  scale_y_continuous(
    limits = c(0, my_mult), labels = comma_format(.1), breaks = seq(0, 2.5, .5)
  ) +
  scale_x_continuous(breaks = seq(1950, 2030, 10)) +
  scale_fill_manual(
    values = as.character(pal$old[c("plate", "dust")]),
    labels = c("caught steeling", "swiped")
  ) +
  labs(
    title = str_wrap("Stolen bases way up thanks to rule changes this year.", 28),
    subtitle = glue("Stolen bases per nine innings."),
    caption = "Source: MLB Stats API.",
    x = NULL, y = NULL, fill = NULL
  )
```

<font size=-1 color="#666666"><b>Fig. `r (fig_no <- fig_no + 1)`. Stolen base attempts and success rate</b>. Attempts and success rate are up in 2023, but attempts are still well below modern era max in 1987. <br><b>Source</b>: MLB Stats API.</font>

# Appendix D: Attendance

I'm curious about attendance at Cleveland baseball games. Attendance data is spotty prior to 1940 (Fig. `r fig_no + 1`), so perhaps we should take the season averages with a grain of salt for those years.

```{r}
bref_cle_season_ts %>%
  ggplot(aes(x = season, y = na_home_att)) + 
  geom_col(width = 1.0, fill = pal$cle["R"], alpha = 0.6) +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  annotate(
    geom = "curve", x = 2012, y = 35, xend = 2017, yend = 25, 
    curvature = .10, arrow = arrow(length = unit(2, "mm")),
    color = pal$cle["B"]
  ) +
  annotate(
    geom = "text", x = 2007, y = 38, label = "COVID", size = 2.5, 
    family = "Tahoma", hjust = "left", color = "#0C2340") +
  coord_cartesian(ylim = c(0, 80)) +
  labs(
    title = "Cleveland attendance data is not reliable until 1937.",
    subtitle = "Games with no recorded attendance.",
    x = NULL, y = NULL
  )
```

<font size=-1 color="#666666"><b>Fig. `r (fig_no <- fig_no + 1)`</b>. Cleveland home games with no recorded attendance. Attendance at double-headers is only recorded for one game, so missing attendance related to double-headers is filtered out of this figure.<br>
Source: www.baseball-reference.com.</font>

Attendance is constrained by the park capacity. Attendance might have been higher, at least in the 1990s when Cleveland sold out 455 straight games from 6/12/1995 through 4/2/2001. Here is a brief look at capacity:

- [League Park](https://en.wikipedia.org/wiki/League_Park) was built in 1891 and seated 9,000. The Spiders played their until folding after the 1899 season. The Indians (originally the Blues) started in 1901. In 1910 the park was rebuilt with concrete and steel with 18,000 seats.
- Municipal Stadium opened in 1932. The Indians played there from July 1932 through 1933, but returned to League Park in 1934 because attendance didn't justify such a large facility. Beginning in 1937, the Indians played weekend and holiday games at Municipal stadium and gradually added more dates. Beginning in 1947 all games were played there.
- [Progressive Field](https://en.wikipedia.org/wiki/Progressive_Field) (named Jacobs Field from 1994-2007) opened in 1994 with a capacity of 42,865. Modifications increased capacity to 45,199 in 2009. Other modifications in 2011 reduced capacity to 43,441, then 36,856 in 2015, and 34,830 in 2021.

Attendance is generally correlated with winning, but something seems wrong with the two World Series years, 1948 and 1954. Average attendance in 1947, the first full season at Municipal Stadium, was 25,300. The Tribe finished 80-74 that year. The next season the Tribe went 97-58 and went on to win the World Series. Average attendance swelled to 40,300. Despite finishing 89-65 in 1949, attendance fell to 36,300, then to 26,700 in 1950. It continued to fall each season, reaching a low of 16,800 in 1953, despite 90+ wins each year. The 1954 season, with the gaudy 111-43 record only lifted attendance to 20,000 per game, attendance resumed its decline the following season. It briefly rebounded to 23,800 in 1959, but would not exceed 20,000 again until 1993.

```{r}
SF <- 50000
(step_2_fig1 <- bref_cle_season_ts %>%
  mutate(attend = if_else(season == 2020, 0, home_att_G)) %>%
  ggplot(aes(x = season, y = attend)) +
  geom_col(aes(y = 1 * SF), fill = pal$cle["B"], width = 1, alpha = 0.05) +
  geom_col(aes(y = W / (W + L) * SF), fill = pal$cle["B"], width = 1, alpha = 0.2) +
  geom_line(color = pal$cle["R"], na.rm = TRUE, linewidth = 1) +
  geom_point(aes(y = if_else(season %in% c(1920, 1948, 1954, 1995, 1997, 2016), attend, NA_real_)),
             shape = 21, color = pal$cle["R"], fill = "white", size = 3, na.rm = TRUE) +
  geom_hline(yintercept = 0.50 * SF, linetype = 3, linewidth = 1, color = pal$cle["B"]) +
  annotate(
    geom = "curve", x = 1960, y = 28000, xend = 1954, yend = 22000, 
    curvature = .3, arrow = arrow(length = unit(2, "mm")),
    color = pal$cle["B"]
  ) +
  annotate("text", x = 1961, y = 28000, label = "1954", size = 3, 
           hjust = "left", color = pal$cle["B"]) +
  annotate(
    geom = "curve", x = 2012, y = 3000, xend = 2017, yend = 2000, 
    curvature = .10, arrow = arrow(length = unit(2, "mm")),
    color = pal$cle["B"]
  ) +
  annotate(geom = "text", x = 2007, y = 5000, label = "COVID", size = 2.5, 
           hjust = "left", color = pal$cle["B"]) +
  scale_y_continuous(limits = c(0, NA), 
                     labels = scales::comma,
                     expand = c(0,0),
                     sec.axis = sec_axis(trans = ~ .x / SF,
                                         labels = scales::percent_format(),
                                         name = "Winning Percentage")) +
  scale_x_continuous(breaks = seq(1900, 2030, 10), expand = c(0,0)) +
  labs(
    x = NULL, y = "Fans per Game",
    title = "Attendance usually increases with performance.",
    subtitle = "Cleveland average home attendance and winning percentage."
  ) +
  theme(
    axis.text.y.left = element_text(color = pal$cle["R"]),
    axis.title.y.left = element_text(color = pal$cle["R"], margin = margin(r = 5)),
    axis.title.y.right = element_text(color = pal$cle["B"], margin = margin(l = 5))
  )
)
```

<font size=-1 color="#666666"><b>Fig. `r (fig_no <- fig_no + 1)`</b>. Cleveland average home attendance and winning percentage</b>. Attendance is generally accompanies improved play. World Series appearances years are highlighted. 1948 and 1954 stand out. 1948 is incredibly high relative to other winning seasons, and 1954 is low considering the dominant performance. 2023 home attendance YTD averages `r bref_cle_season_ts %>% filter(season == 2023) %>% pull(home_att_G) %>% comma(1)`.<br>Source: www.baseball-reference.com.</font>

As an aside, check out 1917 and 1918. Game attendance does not seem unusual those two years despite WWI and the influenza pandemic. The US entered WWI in Apr '17 and the war ended after the season concluded in '18. The 1918 influenza pandemic started in the spring, but seems to have really taken off around October. Baseball survived the war as an acceptable diversion from civic duty, and in fact patriotic elements were integrated into the games.

For perspective, here is the Major League average with Cleveland attendance in the backdrop. MLB attendance trended upward until about 2007. The drop in 1995 followed the 1994-5 player strike. It seems possible all of MLB received a post-WWII attendance boost that corrected and returned to trend by the early 1950s.

```{r}
x <- bind_rows(
  MLB = retro_season_g_ts %>% replace_na(list(att_G = 0)) %>% as_tibble(),
  Cleveland = bref_cle_season_ts %>% replace_na(list(home_att_G = 0)) %>% 
    rename(att_G = home_att_G) %>% as_tibble(),
  .id = "grp"
)

x %>% 
  ggplot(aes(x = season, y = att_G)) +
  geom_line(aes(color = grp, linewidth = grp, alpha = grp)) +
  geom_point(data = x %>% filter((grp == "MLB" & season %in% c(1995, 2007)) |
                                   (grp == "Cleveland" & season %in% c(1999))),
             aes(color = grp),
             shape = 21, fill = "white", size = 3, na.rm = TRUE) +
  geom_text(data = x %>% filter((grp == "MLB" & season %in% c(1995, 2007)) |
                                   (grp == "Cleveland" & season %in% c(1999))),
            aes(label = glue("{season}: {comma(round(att_G, -2), 1)}")),
            color = pal$sfg["B"], size = 3, hjust = 0, nudge_x = 2, nudge_y = 1000) +
  annotate(
    geom = "curve", x = 2012, y = 3000, xend = 2017, yend = 2000,
    curvature = .10, arrow = arrow(length = unit(2, "mm")),
    color = pal$sfg["B"]
  ) +
  annotate(geom = "text", x = 2007, y = 5000, label = "COVID", size = 2.5,
           hjust = "left", color = pal$sfg["B"]) +
  scale_y_continuous(limits = c(0, 45000), labels = scales::comma, expand = c(0,0)) +
  scale_x_continuous(breaks = seq(1900, 2030, 10), expand = c(0,0)) +
  scale_color_manual(values = as.character(c(pal$sfg["O"], pal$cle["R"]))) +
  scale_linewidth_manual(values = c(1, 2.5)) +
  scale_alpha_manual(values = c(1, .20)) +
  labs(
    x = NULL, y = "Fans per Game", color = NULL, alpha = NULL, linewidth = NULL,
    title = "Cleveland attendance compared to rest of MLB.",
    subtitle = "Average Cleveland attendance. MLB shown in backdrop."
  )
```

<font size=-1 color="#666666"><b>Fig. `r (fig_no <- fig_no + 1)`</b>. Cleveland attendance peaked in 1999 and was `r bref_cle_season_ts %>% filter(season == 2022) %>% pull(home_att_G) %>% comma(1)` in 2022. 2023 through `r max(bref_cle_games$game_date) %>% format("%b %d")` is `r bref_cle_season_ts %>% filter(season == 2023) %>% pull(home_att_G) %>% comma(1)`. Major League Baseball attendance peaked in 2007. 2022 attendance averaged `r retro_season_g_ts %>% filter(season == 2022) %>% pull(att_G) %>% comma(1)`.<br>Source: MLB API, Retrosheet.</font>

# Appendix E: Other Clevaland Trends

## Doubleheaders {.tabset}

```{r}
double_header <- bref_cle_games_ts %>%
  mutate(doubleheader_ind = if_else(doubleheader_game == 0, "Single Game", "Doubleheader")) %>%
  index_by(season = ~ year(.)) %>%
  group_by(season, doubleheader_ind) %>%
  summarize(hrs_9I = mean(hrs_9I, na.rm = TRUE), n = n())

double_header_effect <- double_header %>%
  as_tibble() %>%
  pivot_wider(id_cols = season, names_from = doubleheader_ind, values_from = hrs_9I) %>%
  mutate(twin_diff = `Doubleheader` - `Single Game`)
```

Prior to 1942, a few doubleheaders would be scheduled each season. There doesn't appear to be a pattern: it might be on a Tuesday, for instance. Sunday doubleheaders became common in 1942, either scheduled or as make-ups. In 1943, most Sundays were scheduled doubleheaders and over 50% of baseball games overall were part of a doubleheader. Scheduled doubleheaders, esp. Sundays, were common through 1978. After that, only two or three might be scheduled.

Doubleheaders have become less common while game times have increased. Is there a negative correlation, perhaps a compensating effect? The answer is no, doubleheaders are irrelevant. The average doubleheader game is shorter by only `r scales::comma(mean(double_header_effect$twin_diff, na.rm = TRUE) * 60 * (-1), accuracy = 1)` minutes.

```{r}
SF <- 50 # scaling factor for sec.axis.
double_header %>%
  ggplot(aes(x = season)) +
  geom_col(aes(y = n / SF, fill = fct_rev(doubleheader_ind)), alpha = 0.2, width = 1.0) +
  geom_line(aes(y = hrs_9I, color = fct_rev(doubleheader_ind)), linewidth = 1) +
  scale_color_manual(values = as.character(pal$cle)) +
  scale_fill_manual(values = as.character(pal$cle)) +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  scale_y_continuous(limits = c(0, NA), n.breaks = 4, labels = pretty_hm,
    sec.axis = sec_axis(trans = ~ .x * SF, labels = comma_format(), name = "Games Played")
  ) +
  coord_cartesian(xlim = c(1900, 2030), ylim = c(0, 3.5)) +
  labs(x = NULL, y = "Hours", color = NULL, fill = NULL,
       titletitle = glue("Doubleheader games are shorter by only {scales::comma(mean(double_header_effect$twin_diff, na.rm = TRUE) * 60 * (-1), accuracy = 1)} minutes on average."),
       subtitle = "Cleveland doubleheader game duration.",
       caption = "source: www.baseball-reference.com.") +
  theme(
    legend.position = "top",
    axis.title.y.right = element_text(margin = margin(l = 5))
  )
```

## Night Games

```{r}
day_night <- bref_cle_games_ts %>%
  index_by(season = ~ year(.)) %>%
  group_by(season, day_ind) %>%
  summarize(hrs_9I = mean(hrs_9I, na.rm = TRUE), n = n())
```

Until the late 1930s, all games were played in the afternoon. By the mid 1970s day games were primarily on weekends. While there might be a different dynamic to paying in the afternoon (*it's hot - let's get out of here!*), the data doesn't bear that out. The advent of night games is not associated with longer games.

```{r}
SF <- 50 # scaling factor for sec.axis.
day_night %>%
  ggplot(aes(x = season)) +
  geom_col(aes(y = n / SF, fill = day_ind), alpha = 0.2, width = 1.0) +
  geom_line(aes(y = hrs_9I, color = day_ind), linewidth = 1) +
  scale_color_manual(values = as.character(pal$cle)) +
  scale_fill_manual(values = as.character(pal$cle)) +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  scale_y_continuous(limits = c(0, NA), n.breaks = 4, labels = pretty_hm,
    sec.axis = sec_axis(trans = ~ .x * SF, labels = comma_format(), name = "Games Played")
  ) +
  coord_cartesian(xlim = c(1900, 2030), ylim = c(0, 3.5)) +
  labs(x = NULL, y = "Hours", color = NULL, fill = NULL,
       title = "Night games are more common, but they're not longer.",
       subtitle = "Cleveland Indians average game duration, with periods of growth and stability.",
       caption = "source: www.baseball-reference.com.") +
  theme(
    legend.position = "top",
    axis.title.y.right = element_text(margin = margin(l = 5))
  )
```

# Appendix F: Other Notes

According to [this Baseball Reference article on Pace of Play](https://www.baseball-reference.com/bullpen/Pace_of_play), Bill James identified mid-inning pitching substitutions as a break in the momentum of the game. James might appears to have been more concerned with the entertainment quality rather than pace. It also lists measures proposed to hasten pace, many of which are now in place: automatic intentional walks, regulating stepping out of batters box, pitch clock, shorter time between innings, limited mound visits, the three-batter rule, mound visit clock, and pitching change clock. The article references David Smith's 2018 article (see below) attributing increased times to deeper counts and strikeouts leading to higher pitch counts. The pitch clock, introduced in the minors in 2022 reduced average game time by 25 minutes.

David W. Smith: "Why Do Games Take So Long?", Baseball Research Journal, SABR, Vol. 47, Nr. 2 (Fall 2018), pp. 64-69. [html](https://sabr.org/journal/article/why-do-games-take-so-long/).].
David Smith (2018) concluded a rising pitch count is the predominant driver of game time. Smith relied on [Retrosheet](https://www.retrosheet.org/) for most of his analysis, but used other data for pitch counts.^[I used Retrosheet data in [my study on baseball player longevity](https://mpfoley73.github.io/baseball-survival/1_full_analysis.html)].

David W. Smith: "Time Between Pitches: Cause of Long Games?", Baseball Research Journal, SABR, Vol. 48 Nr. 2 (Fall 2019), pp. 24-28. [html](https://sabr.org/journal/article/time-between-pitches-cause-of-long-games/)

https://www.nyseconomicsassociation.org/wp-content/uploads/2022/06/nyer-sports-2020.pdf#page=24
https://sabr.org/journal/article/why-do-games-take-so-long/
https://github.com/BillPetti/baseballr/blob/master/R/retrosheet_data.R
https://retrosheet.org/Research/SmithD/Relief%20Pitchers%20from%20Exile%20to%20Specialist.pdf
https://fivethirtyeight.com/features/relievers-have-broken-baseball-we-have-a-plan-to-fix-it/
https://www.pitcherlist.com/retrosheet-play-by-play-data-at-your-fingertips/
https://techgraphs.fangraphs.com/tag/retrosheet/
