==> src/03a_linear_forcast.Rmd <==
---
title: "Cleveland Indians Game Duration"
subtitle: "Time Series Regression"
author: "Michael Foley"
date: "`r Sys.Date()`"
output: 
  html_document:
    css: "style.css"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)

library(tidyverse)
library(lubridate)
library(ggthemes)
library(extrafont)
library(fpp3)
library(tsibble)
library(fable)
library(patchwork) # for arranging plots
library(glue)

source("00_setup.R")
```

```{r data}
# Get data from prior step
cle_tsibble_yr <- readRDS("../data/cle_tsibble_yr.rds")
```

A time series regression forecasts a time series as a linear relationship with the independent variables.

$$y_t = X_t \beta + \epsilon_t$$

The linear regression model assumes there is a **linear** relationship between the forecast variable game time per nine innings (`game_hrs9`) and the predictor variables. For a time series regression the year (`yr`) will be the main predictor variable. It might also make sense to include other predictors, like runs per nine-innings (`runs9`).

```{r}
cle_tsibble_yr %>%
  ggplot(aes(x = yr)) +
  geom_line(aes(y = game_hrs9), size = 1, alpha = 0.6, color = cle_colors["red"]) +
  geom_line(aes(y = runs9 / 5), linetype = 2, color = cle_colors["navy_blue"]) +
  scale_y_continuous(limits = c(0, NA), sec.axis = sec_axis(~ . * 5, name = "Runs")) +
  scale_x_continuous(breaks = seq(1900, 2030, 10)) +
  theme(axis.text.y.left = element_text(color = cle_colors["red"]),
        axis.title.y.left = element_text(color = cle_colors["red"])) +
  labs(x = NULL, y = "Hours", color = NULL, fill = NULL,
       title = "Game duration updward trend with possible relationship to scoring.",
       subtitle = "Cleveland Indians average game duration, runs per game.",
       caption = "source: www.baseball-reference.com.")
```

The correlation matrix shows `game_hrs9` is positively correlated with `yr`, but only mildly correlated with `runs9`. 

```{r}
cle_tsibble_yr %>%
  as_tibble() %>%
  filter(!is.na(game_hrs9)) %>%
  select(-c(home_att, wins, losses)) %>%
  cor() %>%
  ggcorrplot::ggcorrplot(type = "upper", lab = TRUE, lab_size = 3) +
  labs(title = "Game duration is correlated with year, but not scoring.",
       subtitle = "Correlation plot of Tribe seasonal statistics.", 
       caption = "Source: www.baseball-reference.com.",
       x = NULL, y = NULL)
```

## Fit Model

A good first start is a simple regression on `yr` (or just the trend). 

$$\mathrm{DURATION}_t = \beta_0 + \beta_1 \mathrm{YR}_t + \epsilon_t$$

```{r warning=FALSE}
fmla <- formula(game_hrs9 ~ yr)

mdl_reg <- cle_tsibble_yr %>%
  model(TSLM(fmla))

mdl_reg %>% report()
```

The modeled $R^2$ is `r mdl_reg %>% broom::glance() %>% pull("r_squared") %>% scales::comma(accuracy = .001)`, the adjusted $R^2$ is `r mdl_reg %>% broom::glance() %>% pull("adj_r_squared") %>% scales::comma(accuracy = .001)`, and the standard error of the regression, $\hat{\sigma}_\epsilon,$ is `r mdl_reg %>% broom::glance() %>% pull(sigma2) %>% sqrt() %>% scales::comma(accuracy = .01)` (`r mdl_reg %>% broom::glance() %>% pull(sigma2) %>% sqrt() %>% prod(60) %>% scales::comma(accuracy = 1)` minutes). The fitted to actuals plot has a good linear relationship.

```{r}
p1 <- augment(mdl_reg) %>%
  ggplot(aes(x = yr)) +
  geom_line(aes(y = game_hrs9), color = "dark gray", size = 1) +
  geom_line(aes(y = .fitted), color = "goldenrod", size = 1) +
  labs(subtitle = "Time series")
p2 <- augment(mdl_reg) %>%
  ggplot(aes(x = game_hrs9, y = .fitted)) +
  geom_point(color = "goldenrod", size = 1) +
  geom_abline(intercept = 0, slope = 1, linetype = 2, size = 1, color = "dark gray") +
  labs(subtitle = "Fitted vs actuals")
  
p1 + p2 +
  patchwork::plot_annotation(title = "Fitted values plots",
                             subtitle = fmla)
```

## Model Evaluation

Evaluate the regression model with diagnostic plots. Use `feasts::gg_tsresiduals()` from the tidyverts.

```{r}
gg_tsresiduals(mdl_reg)
```

Time series observations are usually related to prior observations. That shows up in diagnostic plots as autocorrelation in the residuals. Autocorrelation in the residuals increases the prediction intervals, making forecasts less efficient (although still unbiased). The autocorrelation function plot (ACF) finds a significant positive spike at lags 1-3 and a negative spike at lag 20. 

The residuals vs time diagnostic plot revealed no heteroscedasticity. Heteroscedasticity can make prediction intervals inaccurate.

The histogram shows that the residuals are slightly right-skewed. Non-normality of the residuals can also make the prediction intervals inaccurate.

The residuals should be independent of each of the explanatory variables *and* independent of candidate variables not used in the model. In this case, the residuals have a random pattern in each of the plots.

```{r}
cle_tsibble_yr %>%
  left_join(residuals(mdl_reg), by = "yr") %>% 
  pivot_longer(cols = c(runs9, wins, home_att), names_to = "regressor", values_to = "x") %>%
  ggplot(aes(x = x, y = .resid, color = regressor)) +
  geom_point(show.legend = FALSE) +
  facet_wrap(vars(regressor), scales = "free_x") +
  labs(title = "There is no relationship between residuals and individual regressors.",
       subtitle = "otherwise the relationship may be nonlinear.",
       x = NULL) +
  ggthemes::scale_color_few()
```

A second check on the homoscedastity assumption is a plot of the residuals against the fitted values. Again, there should be no pattern. This one shows a cyclic pattern.

```{r}
augment(mdl_reg) %>%
  ggplot(aes(x = .fitted, y = .resid)) +
  geom_point() +
  labs(title = "There is cyclic relationship between residuals and fitted values.",
       subtitle = "The response variable may require transformation.",
       y = "Residuals", x = "Fitted")
```

Let's try a non-linear option. If there are kinks in the response variable trend, you can use piecewise linear regression by defining the knots in the series. In the exploration phase, I identified knots at 1920, 1945, 1955, 1975, 2000, and 2003.

```{r}
fmla_pls <- formula(game_hrs9 ~ trend(knots = c(1920, 1945, 1955, 1975, 2000, 2003)))

mdl_reg_pls <- cle_tsibble_yr %>%
  model(TSLM(fmla_pls))

mdl_reg_pls %>% report()
```

```{r}
augment(mdl_reg_pls) %>%
  ggplot(aes(x = .fitted, y = .resid)) +
  geom_point() +
  labs(title = "There is no relationship between residuals and fitted values.",
       subtitle = "The response variable requires no transformation.",
       y = "Residuals", x = "Fitted")
```

## Predicting Values

Use the model to forecast game durations over the next decade.

```{r warning=FALSE}
fc_reg_pls <- forecast(mdl_reg_pls, h = 10, point_forecast = list(mu = mean))

fc_reg_pls_aug <- fc_reg_pls %>%
  mutate(
    hilo95 = hilo(game_hrs9, .95),
    ci_025 = map_dbl(hilo95, ~.x$lower),
    ci_975 = map_dbl(hilo95, ~.x$upper),
    hilo80 = hilo(game_hrs9, .80),
    ci_10 = map_dbl(hilo80, ~.x$lower),
    ci_90 = map_dbl(hilo80, ~.x$upper)
  ) %>%
  select(yr, game_hrs9, mu, ci_025, ci_975, ci_10, ci_90)

mdl_reg_pls %>% 
  augment() %>%
  ggplot(aes(x = yr)) +
  geom_line(aes(y = game_hrs9), color = "dark gray") +
  geom_line(aes(y = .fitted), color = "goldenrod") +
  geom_line(data = fc_reg_pls_aug, aes(y = mu), color = "goldenrod", size = 1) +
  geom_ribbon(data = fc_reg_pls_aug, 
              aes(ymin = ci_025, ymax = ci_975), 
              fill = "goldenrod", alpha = .2) +
  scale_y_continuous(limits = c(0, NA)) +
  scale_x_continuous(breaks = seq(1900, 2040, 10)) +
  labs(title = "Piecewise linear model with 10 year forecast.",
       caption = "Shaded area is 95% confidence interval.",
       x = NULL, y = "Hours", color = NULL, fill = NULL)
```

The model predicts game durations to increase to `r fc_reg_pls %>% filter(yr == 2031) %>% pull(mu) %>% pretty_hm()` by 2031.
